stop using the json API at pypi
