+++ tests/installation/fixtures/with-pypi-repository.test
@@ -51,7 +51,7 @@ name = "pluggy"
 version = "0.6.0"
 description = "plugin and hook calling mechanisms for python"
 optional = false
-python-versions = ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*"
+python-versions = ">=2.7,<3.0 || >=3.4"
 groups = ["dev"]
 files = [
     {file = "pluggy-0.6.0-py2-none-any.whl", hash = "sha256:f5f767d398f18aa177976bf9c4d0c05d96487a7d8f07062251585803aaf56246"},
