--- tests/repositories/fixtures/pypi.org/json/attrs/17.4.0.json
@@ -1,121 +1,0 @@
-{
-  "info": {
-    "author": "Hynek Schlawack",
-    "author_email": "hs@ox.cx",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Natural Language :: English",
-      "Operating System :: OS Independent",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: Implementation :: CPython",
-      "Programming Language :: Python :: Implementation :: PyPy",
-      "Topic :: Software Development :: Libraries :: Python Modules"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://www.attrs.org/",
-    "keywords": "class,attribute,boilerplate",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "attrs",
-    "package_url": "https://pypi.org/project/attrs/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/attrs/",
-    "project_urls": {
-      "Homepage": "http://www.attrs.org/"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/attrs/17.4.0/",
-    "requires_dist": [
-      "coverage; extra == 'dev'",
-      "hypothesis; extra == 'dev'",
-      "pympler; extra == 'dev'",
-      "pytest; extra == 'dev'",
-      "six; extra == 'dev'",
-      "zope.interface; extra == 'dev'",
-      "sphinx; extra == 'dev'",
-      "zope.interface; extra == 'dev'",
-      "sphinx; extra == 'docs'",
-      "zope.interface; extra == 'docs'",
-      "coverage; extra == 'tests'",
-      "hypothesis; extra == 'tests'",
-      "pympler; extra == 'tests'",
-      "pytest; extra == 'tests'",
-      "six; extra == 'tests'",
-      "zope.interface; extra == 'tests'"
-    ],
-    "requires_python": "",
-    "summary": "Classes Without Boilerplate",
-    "version": "17.4.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "attrs",
-    "roles": []
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "7fe37931797b16c7fa158017457a9ea9",
-        "sha256": "1fbfc10ebc8c876dcbab17f016b80ae1a4f0c1413461a695871427960795beb4"
-      },
-      "downloads": -1,
-      "filename": "attrs-17.4.0-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "7fe37931797b16c7fa158017457a9ea9",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": null,
-      "size": 31658,
-      "upload_time": "2017-12-30T08:20:05",
-      "upload_time_iso_8601": "2017-12-30T08:20:05.582456Z",
-      "url": "https://files.pythonhosted.org/packages/b5/60/4e178c1e790fd60f1229a9b3cb2f8bc2f4cc6ff2c8838054c142c70b5adc/attrs-17.4.0-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "c03e5b3608d9071fbd098850d8922668",
-        "sha256": "eb7536a1e6928190b3008c5b350bdf9850d619fff212341cd096f87a27a5e564"
-      },
-      "downloads": -1,
-      "filename": "attrs-17.4.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "c03e5b3608d9071fbd098850d8922668",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 97071,
-      "upload_time": "2017-12-30T08:20:08",
-      "upload_time_iso_8601": "2017-12-30T08:20:08.575620Z",
-      "url": "https://files.pythonhosted.org/packages/8b/0b/a06cfcb69d0cb004fde8bc6f0fd192d96d565d1b8aa2829f0f20adb796e5/attrs-17.4.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
