--- tests/repositories/fixtures/pypi.org/json/cleo/1.0.0a5.json
@@ -1,107 +1,0 @@
-{
-  "info": {
-    "author": "Sébastien Eustace",
-    "author_email": "sebastien@eustace.io",
-    "bugtrack_url": null,
-    "classifiers": [
-      "License :: OSI Approved :: MIT License",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.10",
-      "Programming Language :: Python :: 3.7",
-      "Programming Language :: Python :: 3.8",
-      "Programming Language :: Python :: 3.9"
-    ],
-    "description": "",
-    "description_content_type": "text/markdown",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/python-poetry/cleo",
-    "keywords": "cli,commands",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "cleo",
-    "package_url": "https://pypi.org/project/cleo/",
-    "platform": null,
-    "project_url": "https://pypi.org/project/cleo/",
-    "project_urls": {
-      "Homepage": "https://github.com/python-poetry/cleo"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/cleo/1.0.0a5/",
-    "requires_dist": [
-      "pylev (>=1.3.0,<2.0.0)",
-      "crashtest (>=0.3.1,<0.4.0)"
-    ],
-    "requires_python": ">=3.7,<4.0",
-    "summary": "Cleo allows you to create beautiful and testable command-line interfaces.",
-    "version": "1.0.0a5",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "poetry",
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "neersighted"
-      },
-      {
-        "role": "Maintainer",
-        "user": "secrus"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "19ed7de77063e8f16bc459276ccbe197",
-        "sha256": "d0cfea878b77be28be027033e6af419b705abe47278067a7c3a298f39cf825c5"
-      },
-      "downloads": -1,
-      "filename": "cleo-1.0.0a5-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "19ed7de77063e8f16bc459276ccbe197",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": ">=3.7,<4.0",
-      "size": 78701,
-      "upload_time": "2022-06-03T20:16:19",
-      "upload_time_iso_8601": "2022-06-03T20:16:19.386916Z",
-      "url": "https://files.pythonhosted.org/packages/45/0c/3825603bf62f360829b1eea29a43dadce30829067e288170b3bf738aafd0/cleo-1.0.0a5-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "92e181952976e09b9d1c583da6c3e2fc",
-        "sha256": "88f0a4275a17f2ab4d013786b8b9522d4c60bd37d8fc9b3def0fb27f4ac1e694"
-      },
-      "downloads": -1,
-      "filename": "cleo-1.0.0a5.tar.gz",
-      "has_sig": false,
-      "md5_digest": "92e181952976e09b9d1c583da6c3e2fc",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=3.7,<4.0",
-      "size": 61431,
-      "upload_time": "2022-06-03T20:16:21",
-      "upload_time_iso_8601": "2022-06-03T20:16:21.133890Z",
-      "url": "https://files.pythonhosted.org/packages/2f/16/1c1902b225756745f9860451a44a2e2a3c26ee91c72295e83c63df605ed1/cleo-1.0.0a5.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
