--- tests/repositories/fixtures/pypi.org/json/clikit/0.2.4.json
@@ -1,108 +1,0 @@
-{
-  "info": {
-    "author": "Sébastien Eustace",
-    "author_email": "sebastien@eustace.io",
-    "bugtrack_url": null,
-    "classifiers": [
-      "License :: OSI Approved :: MIT License",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: 3.7"
-    ],
-    "description": "",
-    "description_content_type": "text/markdown",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/sdispater/clikit",
-    "keywords": "packaging,dependency,poetry",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "Sébastien Eustace",
-    "maintainer_email": "sebastien@eustace.io",
-    "name": "clikit",
-    "package_url": "https://pypi.org/project/clikit/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/clikit/",
-    "project_urls": {
-      "Homepage": "https://github.com/sdispater/clikit",
-      "Repository": "https://github.com/sdispater/clikit"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/clikit/0.2.4/",
-    "requires_dist": [
-      "pastel (>=0.1.0,<0.2.0)",
-      "pylev (>=1.3,<2.0)",
-      "typing (>=3.6,<4.0); python_version >= \"2.7\" and python_version < \"2.8\" or python_version >= \"3.4\" and python_version < \"3.5\"",
-      "enum34 (>=1.1,<2.0); python_version >= \"2.7\" and python_version < \"2.8\""
-    ],
-    "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-    "summary": "CliKit is a group of utilities to build beautiful and testable command line interfaces.",
-    "version": "0.2.4",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "SDisPater"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "93a51e8bf259c29692e51a7cbca6d664",
-        "sha256": "27316bf6382b04be8fb2f60c85d538fd2b2b03f0f1eba5c88f7d7eddbefc2778"
-      },
-      "downloads": -1,
-      "filename": "clikit-0.2.4-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "93a51e8bf259c29692e51a7cbca6d664",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 85786,
-      "upload_time": "2019-05-11T17:09:23",
-      "upload_time_iso_8601": "2019-05-11T17:09:23.516387Z",
-      "url": "https://files.pythonhosted.org/packages/7b/0d/bb4c8a2d0edca8c300373ed736fb4680cf73be5be2ff84544dee5f979c14/clikit-0.2.4-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "f7cdbad3508038a04561f646aae68146",
-        "sha256": "0fdd41e86e8b118a8b1e94ef2835925ada541d481c9b3b2fc635fa68713e6125"
-      },
-      "downloads": -1,
-      "filename": "clikit-0.2.4.tar.gz",
-      "has_sig": false,
-      "md5_digest": "f7cdbad3508038a04561f646aae68146",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 50980,
-      "upload_time": "2019-05-11T17:09:25",
-      "upload_time_iso_8601": "2019-05-11T17:09:25.865051Z",
-      "url": "https://files.pythonhosted.org/packages/c5/33/14fad4c82f256b0ef60dd25d4b6d8145b463da5274fd9cd842f06af318ed/clikit-0.2.4.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
