--- tests/repositories/fixtures/pypi.org/json/colorama/0.3.9.json
@@ -1,115 +1,0 @@
-{
-  "info": {
-    "author": "Arnon Yaari",
-    "author_email": "tartley@tartley.com",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Environment :: Console",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: BSD License",
-      "Operating System :: OS Independent",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.5",
-      "Programming Language :: Python :: 2.6",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.1",
-      "Programming Language :: Python :: 3.2",
-      "Programming Language :: Python :: 3.3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Topic :: Terminals"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "UNKNOWN",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/tartley/colorama",
-    "keywords": "color colour terminal text ansi windows crossplatform xplatform",
-    "license": "BSD",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": null,
-    "maintainer_email": null,
-    "name": "colorama",
-    "package_url": "https://pypi.org/project/colorama/",
-    "platform": "UNKNOWN",
-    "project_url": "https://pypi.org/project/colorama/",
-    "project_urls": {
-      "Download": "UNKNOWN",
-      "Homepage": "https://github.com/tartley/colorama"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/colorama/0.3.9/",
-    "requires_dist": null,
-    "requires_python": null,
-    "summary": "Cross-platform colored terminal text.",
-    "version": "0.3.9",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "tartley"
-      },
-      {
-        "role": "Owner",
-        "user": "wiggin15"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "8021c861015b5f590be41190bc3f8eed",
-        "sha256": "78a441d2e984c790526cdef1cfd8415a366979ef5b3186771a055b35886953bf"
-      },
-      "downloads": -1,
-      "filename": "colorama-0.3.9-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "8021c861015b5f590be41190bc3f8eed",
-      "packagetype": "bdist_wheel",
-      "python_version": "2.7",
-      "requires_python": null,
-      "size": 20181,
-      "upload_time": "2017-04-27T07:12:36",
-      "upload_time_iso_8601": "2017-04-27T07:12:36.597052Z",
-      "url": "https://files.pythonhosted.org/packages/db/c8/7dcf9dbcb22429512708fe3a547f8b6101c0d02137acbd892505aee57adf/colorama-0.3.9-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "8323a5b84fdf7ad810804e51fc256b39",
-        "sha256": "4c5a15209723ce1330a5c193465fe221098f761e9640d823a2ce7c03f983137f"
-      },
-      "downloads": -1,
-      "filename": "colorama-0.3.9.tar.gz",
-      "has_sig": false,
-      "md5_digest": "8323a5b84fdf7ad810804e51fc256b39",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 25053,
-      "upload_time": "2017-04-27T07:12:12",
-      "upload_time_iso_8601": "2017-04-27T07:12:12.351237Z",
-      "url": "https://files.pythonhosted.org/packages/e6/76/257b53926889e2835355d74fec73d82662100135293e17d382e2b74d1669/colorama-0.3.9.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
