--- tests/repositories/fixtures/pypi.org/json/discord-py/2.0.0.json
@@ -1,126 +1,0 @@
-{
-  "info": {
-    "author": "Rapptz",
-    "author_email": "",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Natural Language :: English",
-      "Operating System :: OS Independent",
-      "Programming Language :: Python :: 3.10",
-      "Programming Language :: Python :: 3.8",
-      "Programming Language :: Python :: 3.9",
-      "Topic :: Internet",
-      "Topic :: Software Development :: Libraries",
-      "Topic :: Software Development :: Libraries :: Python Modules",
-      "Topic :: Utilities",
-      "Typing :: Typed"
-    ],
-    "description": "",
-    "description_content_type": "text/x-rst",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/Rapptz/discord.py",
-    "keywords": "",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "discord.py",
-    "package_url": "https://pypi.org/project/discord.py/",
-    "platform": null,
-    "project_url": "https://pypi.org/project/discord.py/",
-    "project_urls": {
-      "Documentation": "https://discordpy.readthedocs.io/en/latest/",
-      "Homepage": "https://github.com/Rapptz/discord.py",
-      "Issue tracker": "https://github.com/Rapptz/discord.py/issues"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/discord.py/2.0.0/",
-    "requires_dist": [
-      "aiohttp (<4,>=3.7.4)",
-      "sphinx (==4.4.0) ; extra == 'docs'",
-      "sphinxcontrib-trio (==1.1.2) ; extra == 'docs'",
-      "sphinxcontrib-websupport ; extra == 'docs'",
-      "typing-extensions (<5,>=4.3) ; extra == 'docs'",
-      "orjson (>=3.5.4) ; extra == 'speed'",
-      "aiodns (>=1.1) ; extra == 'speed'",
-      "Brotli ; extra == 'speed'",
-      "cchardet ; extra == 'speed'",
-      "coverage[toml] ; extra == 'test'",
-      "pytest ; extra == 'test'",
-      "pytest-asyncio ; extra == 'test'",
-      "pytest-cov ; extra == 'test'",
-      "pytest-mock ; extra == 'test'",
-      "typing-extensions (<5,>=4.3) ; extra == 'test'",
-      "PyNaCl (<1.6,>=1.3.0) ; extra == 'voice'"
-    ],
-    "requires_python": ">=3.8.0",
-    "summary": "A Python wrapper for the Discord API",
-    "version": "2.0.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "rapptz"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "65394fc868632423cedb6be7259db970",
-        "sha256": "25b9739ba456622655203a0925b354c0ba96ac6c740562e7c37791c2f6b594fb"
-      },
-      "downloads": -1,
-      "filename": "discord.py-2.0.0-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "65394fc868632423cedb6be7259db970",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": ">=3.8.0",
-      "size": 1059049,
-      "upload_time": "2022-08-18T03:47:52",
-      "upload_time_iso_8601": "2022-08-18T03:47:52.438785Z",
-      "url": "https://files.pythonhosted.org/packages/0e/d9/7b057cab41c16144925ba4f96dab576a8ebb7b80a98d40e06bd94298eb3b/discord.py-2.0.0-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "6c0505a6032342b29f31f9979f37d277",
-        "sha256": "b86fa9dd562684f7a52564e6dfe0216f6c172a009c0d86b8dea8bdd6ffa6b1f4"
-      },
-      "downloads": -1,
-      "filename": "discord.py-2.0.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "6c0505a6032342b29f31f9979f37d277",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=3.8.0",
-      "size": 955054,
-      "upload_time": "2022-08-18T03:47:54",
-      "upload_time_iso_8601": "2022-08-18T03:47:54.173712Z",
-      "url": "https://files.pythonhosted.org/packages/4c/73/fb89115b07588bf7a46e9eca972b89dd62b5856abd52297fe130b41d9d63/discord.py-2.0.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
