--- tests/repositories/fixtures/pypi.org/json/funcsigs/1.0.2.json
@@ -1,120 +1,0 @@
-{
-  "info": {
-    "author": "Testing Cabal",
-    "author_email": "testing-in-python@lists.idyll.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 4 - Beta",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: Apache Software License",
-      "Operating System :: OS Independent",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.6",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: Implementation :: CPython",
-      "Programming Language :: Python :: Implementation :: PyPy",
-      "Topic :: Software Development :: Libraries :: Python Modules"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "UNKNOWN",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://funcsigs.readthedocs.org",
-    "keywords": null,
-    "license": "ASL",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": null,
-    "maintainer_email": null,
-    "name": "funcsigs",
-    "package_url": "https://pypi.org/project/funcsigs/",
-    "platform": "UNKNOWN",
-    "project_url": "https://pypi.org/project/funcsigs/",
-    "project_urls": {
-      "Download": "UNKNOWN",
-      "Homepage": "http://funcsigs.readthedocs.org"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/funcsigs/1.0.2/",
-    "requires_dist": null,
-    "requires_python": null,
-    "summary": "Python function signatures from PEP362 for Python 2.6, 2.7 and 3.2+",
-    "version": "1.0.2",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "Aaron.Iles"
-      },
-      {
-        "role": "Owner",
-        "user": "Joshua.Harlow"
-      },
-      {
-        "role": "Owner",
-        "user": "lifeless"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "69cbf5be453359271714c01b9bd06126eaf2e368f1fddfff30818754b5ac2328",
-        "md5": "701d58358171f34b6d1197de2923a35a",
-        "sha256": "330cc27ccbf7f1e992e69fef78261dc7c6569012cf397db8d3de0234e6c937ca"
-      },
-      "downloads": -1,
-      "filename": "funcsigs-1.0.2-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "701d58358171f34b6d1197de2923a35a",
-      "packagetype": "bdist_wheel",
-      "python_version": "2.7",
-      "requires_python": null,
-      "size": 17697,
-      "upload_time": "2016-04-25T22:22:05",
-      "upload_time_iso_8601": "2016-04-25T22:22:05.222685Z",
-      "url": "https://files.pythonhosted.org/packages/69/cb/f5be453359271714c01b9bd06126eaf2e368f1fddfff30818754b5ac2328/funcsigs-1.0.2-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "944adb842e7a0545de1cdb0439bb80e6e42dfe82aaeaadd4072f2263a4fbed23",
-        "md5": "7e583285b1fb8a76305d6d68f4ccc14e",
-        "sha256": "a7bb0f2cf3a3fd1ab2732cb49eba4252c2af4240442415b4abce3b87022a8f50"
-      },
-      "downloads": -1,
-      "filename": "funcsigs-1.0.2.tar.gz",
-      "has_sig": false,
-      "md5_digest": "7e583285b1fb8a76305d6d68f4ccc14e",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 27947,
-      "upload_time": "2016-04-25T22:22:33",
-      "upload_time_iso_8601": "2016-04-25T22:22:33.882246Z",
-      "url": "https://files.pythonhosted.org/packages/94/4a/db842e7a0545de1cdb0439bb80e6e42dfe82aaeaadd4072f2263a4fbed23/funcsigs-1.0.2.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
