--- tests/repositories/fixtures/pypi.org/json/futures/3.2.0.json
@@ -1,104 +1,0 @@
-{
-  "info": {
-    "author": "Alex Grönholm",
-    "author_email": "alex.gronholm@nextday.fi",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: Python Software Foundation License",
-      "Programming Language :: Python :: 2 :: Only",
-      "Programming Language :: Python :: 2.6",
-      "Programming Language :: Python :: 2.7"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": "https://pythonhosted.org/futures/",
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/agronholm/pythonfutures",
-    "keywords": "",
-    "license": "PSF",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "futures",
-    "package_url": "https://pypi.org/project/futures/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/futures/",
-    "project_urls": {
-      "Homepage": "https://github.com/agronholm/pythonfutures"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/futures/3.2.0/",
-    "requires_dist": null,
-    "requires_python": ">=2.6, <3",
-    "summary": "Backport of the concurrent.futures package from Python 3",
-    "version": "3.2.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "agronholm"
-      },
-      {
-        "role": "Owner",
-        "user": "bquinlan"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "f81c5c27f3ba2efc008cc96363a81c5e",
-        "sha256": "41353b36198757a766cfc82dc9b60e88ecb28e543dd92473b2cc74fc7bf205af"
-      },
-      "downloads": -1,
-      "filename": "futures-3.2.0-py2-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "f81c5c27f3ba2efc008cc96363a81c5e",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2",
-      "requires_python": ">=2.6, <3",
-      "size": 15847,
-      "upload_time": "2017-11-30T23:22:35",
-      "upload_time_iso_8601": "2017-11-30T23:22:35.590688Z",
-      "url": "https://files.pythonhosted.org/packages/2d/99/b2c4e9d5a30f6471e410a146232b4118e697fa3ffc06d6a65efde84debd0/futures-3.2.0-py2-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "40eb168dab84e606df3fdb7e67fe27b7",
-        "sha256": "baf0d469c9e541b747986b7404cd63a5496955bd0c43a3cc068c449b09b7d4a4"
-      },
-      "downloads": -1,
-      "filename": "futures-3.2.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "40eb168dab84e606df3fdb7e67fe27b7",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=2.6, <3",
-      "size": 27320,
-      "upload_time": "2017-11-30T23:22:36",
-      "upload_time_iso_8601": "2017-11-30T23:22:36.994073Z",
-      "url": "https://files.pythonhosted.org/packages/1f/9e/7b2ff7e965fc654592269f2906ade1c7d705f1bf25b7d469fa153f7d19eb/futures-3.2.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
