--- tests/repositories/fixtures/pypi.org/json/importlib-metadata/1.7.0.json
@@ -1,120 +1,0 @@
-{
-  "info": {
-    "author": "Barry Warsaw",
-    "author_email": "barry@python.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 3 - Alpha",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: Apache Software License",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 3",
-      "Topic :: Software Development :: Libraries"
-    ],
-    "description": "",
-    "description_content_type": "",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://importlib-metadata.readthedocs.io/",
-    "keywords": "",
-    "license": "Apache Software License",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "importlib-metadata",
-    "package_url": "https://pypi.org/project/importlib-metadata/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/importlib-metadata/",
-    "project_urls": {
-      "Homepage": "http://importlib-metadata.readthedocs.io/"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/importlib-metadata/1.7.0/",
-    "requires_dist": [
-      "zipp (>=0.5)",
-      "pathlib2 ; python_version < \"3\"",
-      "contextlib2 ; python_version < \"3\"",
-      "configparser (>=3.5) ; python_version < \"3\"",
-      "sphinx ; extra == 'docs'",
-      "rst.linker ; extra == 'docs'",
-      "packaging ; extra == 'testing'",
-      "pep517 ; extra == 'testing'",
-      "importlib-resources (>=1.3) ; (python_version < \"3.9\") and extra == 'testing'"
-    ],
-    "requires_python": "!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*,>=2.7",
-    "summary": "Read metadata from Python packages",
-    "version": "1.7.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "barry"
-      },
-      {
-        "role": "Owner",
-        "user": "brettcannon"
-      },
-      {
-        "role": "Owner",
-        "user": "jaraco"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "8e58cdea07eb51fc2b906db0968a94700866fc46249bdc75cac23f9d13168929",
-        "md5": "8ae1f31228e29443c08e07501a99d1b8",
-        "sha256": "dc15b2969b4ce36305c51eebe62d418ac7791e9a157911d58bfb1f9ccd8e2070"
-      },
-      "downloads": -1,
-      "filename": "importlib_metadata-1.7.0-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "8ae1f31228e29443c08e07501a99d1b8",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": "!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*,>=2.7",
-      "size": 31809,
-      "upload_time": "2020-06-26T21:38:16",
-      "upload_time_iso_8601": "2020-06-26T21:38:16.079439Z",
-      "url": "https://files.pythonhosted.org/packages/8e/58/cdea07eb51fc2b906db0968a94700866fc46249bdc75cac23f9d13168929/importlib_metadata-1.7.0-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "e2ae0b037584024c1557e537d25482c306cf6327b5a09b6c4b893579292c1c38",
-        "md5": "4505ea85600cca1e693a4f8f5dd27ba8",
-        "sha256": "90bb658cdbbf6d1735b6341ce708fc7024a3e14e99ffdc5783edea9f9b077f83"
-      },
-      "downloads": -1,
-      "filename": "importlib_metadata-1.7.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "4505ea85600cca1e693a4f8f5dd27ba8",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": "!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*,>=2.7",
-      "size": 29233,
-      "upload_time": "2020-06-26T21:38:17",
-      "upload_time_iso_8601": "2020-06-26T21:38:17.338581Z",
-      "url": "https://files.pythonhosted.org/packages/e2/ae/0b037584024c1557e537d25482c306cf6327b5a09b6c4b893579292c1c38/importlib_metadata-1.7.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
