--- tests/repositories/fixtures/pypi.org/json/ipython/4.1.0rc1.json
@@ -1,187 +1,0 @@
-{
-  "info": {
-    "author": "The IPython Development Team",
-    "author_email": "ipython-dev@scipy.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Framework :: IPython",
-      "Intended Audience :: Developers",
-      "Intended Audience :: Science/Research",
-      "License :: OSI Approved :: BSD License",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Topic :: System :: Shells"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "https://github.com/ipython/ipython/downloads",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://ipython.org",
-    "keywords": "Interactive,Interpreter,Shell,Parallel,Distributed,Web-based computing,Qt console,Embedding",
-    "license": "BSD",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "ipython",
-    "package_url": "https://pypi.org/project/ipython/",
-    "platform": "Linux,Mac OSX,Windows XP/Vista/7/8",
-    "project_url": "https://pypi.org/project/ipython/",
-    "project_urls": {
-      "Download": "https://github.com/ipython/ipython/downloads",
-      "Homepage": "http://ipython.org"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/ipython/4.1.0rc1/",
-    "requires_dist": [
-      "pickleshare",
-      "setuptools (>=18.5decorator)",
-      "simplegeneric (>0.8)",
-      "traitlets",
-      "pexpect; sys_platform != \"win32\"",
-      "appnope; sys_platform == \"darwin\"",
-      "gnureadline; sys_platform == \"darwin\" and platform_python_implementation == \"CPython\"",
-      "Sphinx (>=1.3); extra == 'all'",
-      "ipykernel; extra == 'all'",
-      "ipyparallel; extra == 'all'",
-      "ipywidgets; extra == 'all'",
-      "nbconvert; extra == 'all'",
-      "nbformat; extra == 'all'",
-      "nose (>=0.10.1); extra == 'all'",
-      "notebook; extra == 'all'",
-      "qtconsole; extra == 'all'",
-      "requests; extra == 'all'",
-      "testpath; extra == 'all'",
-      "Sphinx (>=1.3); extra == 'doc'",
-      "ipykernel; extra == 'kernel'",
-      "nbconvert; extra == 'nbconvert'",
-      "nbformat; extra == 'nbformat'",
-      "ipywidgets; extra == 'notebook'",
-      "notebook; extra == 'notebook'",
-      "ipyparallel; extra == 'parallel'",
-      "qtconsole; extra == 'qtconsole'",
-      "pyreadline (>=2); sys_platform == \"win32\" and extra == 'terminal'",
-      "nose (>=0.10.1); extra == 'test'",
-      "requests; extra == 'test'",
-      "testpath; extra == 'test'",
-      "mock; python_version == \"2.7\" and extra == 'test'"
-    ],
-    "requires_python": "",
-    "summary": "IPython: Productive Interactive Computing",
-    "version": "4.1.0rc1",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "jupyter",
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "bgranger"
-      },
-      {
-        "role": "Owner",
-        "user": "fperez"
-      },
-      {
-        "role": "Owner",
-        "user": "ivanov"
-      },
-      {
-        "role": "Owner",
-        "user": "minrk"
-      },
-      {
-        "role": "Owner",
-        "user": "takowl"
-      },
-      {
-        "role": "Maintainer",
-        "user": "krassowski"
-      },
-      {
-        "role": "Maintainer",
-        "user": "mbussonn"
-      },
-      {
-        "role": "Maintainer",
-        "user": "trallard"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "ac0204a5d372b4e64f9c97b2846646aec1ce4532885005aa4ba51eb20b80e17f",
-        "md5": "512f0431c850c75a12baa9f8c4a9f12f",
-        "sha256": "4d0a08f3fd8837502bf33e9497a5ab28fe63e2fa4201765f378cb139c7a60d5f"
-      },
-      "downloads": -1,
-      "filename": "ipython-4.1.0rc1-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "512f0431c850c75a12baa9f8c4a9f12f",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": null,
-      "size": 736900,
-      "upload_time": "2016-01-26T19:58:35",
-      "upload_time_iso_8601": "2016-01-26T19:58:35.544443Z",
-      "url": "https://files.pythonhosted.org/packages/ac/02/04a5d372b4e64f9c97b2846646aec1ce4532885005aa4ba51eb20b80e17f/ipython-4.1.0rc1-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "a0def71f0c8b8a26ef28cc968fbf1859729ad3e68146cd2eb4a759e9c88da218",
-        "md5": "2aff56d8e78341f64663bcbc81366376",
-        "sha256": "6244a8e3293088ee31c1854abe1a1e7a409cf3ac2fb7579aa9616bdfadd3d4dc"
-      },
-      "downloads": -1,
-      "filename": "ipython-4.1.0rc1.tar.gz",
-      "has_sig": false,
-      "md5_digest": "2aff56d8e78341f64663bcbc81366376",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 4933377,
-      "upload_time": "2016-01-26T19:58:53",
-      "upload_time_iso_8601": "2016-01-26T19:58:53.938491Z",
-      "url": "https://files.pythonhosted.org/packages/a0/de/f71f0c8b8a26ef28cc968fbf1859729ad3e68146cd2eb4a759e9c88da218/ipython-4.1.0rc1.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "71f09d670266b840b8b921dc7106ecddd892f6fb893424883498e1ba3ec3a3a1",
-        "md5": "a9ff233f176dd99b076b81dc8904ab7a",
-        "sha256": "efa3a5a676648cb18e2a2d3cd6353f3c83f0f704df8eb0eb6ae7d0dcbf187ea1"
-      },
-      "downloads": -1,
-      "filename": "ipython-4.1.0rc1.zip",
-      "has_sig": false,
-      "md5_digest": "a9ff233f176dd99b076b81dc8904ab7a",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 5100723,
-      "upload_time": "2016-01-26T19:59:14",
-      "upload_time_iso_8601": "2016-01-26T19:59:14.449245Z",
-      "url": "https://files.pythonhosted.org/packages/71/f0/9d670266b840b8b921dc7106ecddd892f6fb893424883498e1ba3ec3a3a1/ipython-4.1.0rc1.zip",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
