--- tests/repositories/fixtures/pypi.org/json/ipython/5.7.0.json
@@ -1,193 +1,0 @@
-{
-  "info": {
-    "author": "The IPython Development Team",
-    "author_email": "ipython-dev@python.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Framework :: IPython",
-      "Intended Audience :: Developers",
-      "Intended Audience :: Science/Research",
-      "License :: OSI Approved :: BSD License",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Topic :: System :: Shells"
-    ],
-    "description": "",
-    "description_content_type": "",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://ipython.org",
-    "keywords": "Interactive,Interpreter,Shell,Embedding",
-    "license": "BSD",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "ipython",
-    "package_url": "https://pypi.org/project/ipython/",
-    "platform": "Linux",
-    "project_url": "https://pypi.org/project/ipython/",
-    "project_urls": {
-      "Homepage": "https://ipython.org"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/ipython/5.7.0/",
-    "requires_dist": [
-      "setuptools (>=18.5)",
-      "decorator",
-      "pickleshare",
-      "simplegeneric (>0.8)",
-      "traitlets (>=4.2)",
-      "prompt-toolkit (<2.0.0,>=1.0.4)",
-      "pygments",
-      "backports.shutil-get-terminal-size; python_version == \"2.7\"",
-      "pathlib2; python_version == \"2.7\" or python_version == \"3.3\"",
-      "pexpect; sys_platform != \"win32\"",
-      "appnope; sys_platform == \"darwin\"",
-      "colorama; sys_platform == \"win32\"",
-      "win-unicode-console (>=0.5); sys_platform == \"win32\" and python_version < \"3.6\"",
-      "nbformat; extra == 'all'",
-      "ipykernel; extra == 'all'",
-      "pygments; extra == 'all'",
-      "testpath; extra == 'all'",
-      "notebook; extra == 'all'",
-      "nbconvert; extra == 'all'",
-      "ipyparallel; extra == 'all'",
-      "qtconsole; extra == 'all'",
-      "Sphinx (>=1.3); extra == 'all'",
-      "requests; extra == 'all'",
-      "nose (>=0.10.1); extra == 'all'",
-      "ipywidgets; extra == 'all'",
-      "Sphinx (>=1.3); extra == 'doc'",
-      "ipykernel; extra == 'kernel'",
-      "nbconvert; extra == 'nbconvert'",
-      "nbformat; extra == 'nbformat'",
-      "notebook; extra == 'notebook'",
-      "ipywidgets; extra == 'notebook'",
-      "ipyparallel; extra == 'parallel'",
-      "qtconsole; extra == 'qtconsole'",
-      "nose (>=0.10.1); extra == 'test'",
-      "requests; extra == 'test'",
-      "testpath; extra == 'test'",
-      "pygments; extra == 'test'",
-      "nbformat; extra == 'test'",
-      "ipykernel; extra == 'test'",
-      "mock; python_version == \"2.7\" and extra == 'test'",
-      "numpy; python_version >= \"3.4\" and extra == 'test'"
-    ],
-    "requires_python": "",
-    "summary": "IPython: Productive Interactive Computing",
-    "version": "5.7.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "jupyter",
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "bgranger"
-      },
-      {
-        "role": "Owner",
-        "user": "fperez"
-      },
-      {
-        "role": "Owner",
-        "user": "ivanov"
-      },
-      {
-        "role": "Owner",
-        "user": "minrk"
-      },
-      {
-        "role": "Owner",
-        "user": "takowl"
-      },
-      {
-        "role": "Maintainer",
-        "user": "krassowski"
-      },
-      {
-        "role": "Maintainer",
-        "user": "mbussonn"
-      },
-      {
-        "role": "Maintainer",
-        "user": "trallard"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "20da5e0b1f79dccb37f033a885d798d7",
-        "sha256": "4608e3e0500fe8142659d149891400fc0b9fa250051814b569457ae4688943dc"
-      },
-      "downloads": -1,
-      "filename": "ipython-5.7.0-py2-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "20da5e0b1f79dccb37f033a885d798d7",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2",
-      "requires_python": null,
-      "size": 760413,
-      "upload_time": "2018-05-10T18:56:05",
-      "upload_time_iso_8601": "2018-05-10T18:56:05.083695Z",
-      "url": "https://files.pythonhosted.org/packages/52/19/aadde98d6bde1667d0bf431fb2d22451f880aaa373e0a241c7e7cb5815a0/ipython-5.7.0-py2-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "2844fa01618fe27ab99ad455d605b47d",
-        "sha256": "4292c026552a77b2edc0543941516eddd6fe1a4b681a76ac40b3f585d2fca76f"
-      },
-      "downloads": -1,
-      "filename": "ipython-5.7.0-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "2844fa01618fe27ab99ad455d605b47d",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": null,
-      "size": 760415,
-      "upload_time": "2018-05-10T18:56:07",
-      "upload_time_iso_8601": "2018-05-10T18:56:07.665559Z",
-      "url": "https://files.pythonhosted.org/packages/c7/b6/03e0b5b0972e6161d16c4cec8d41a20372bd0634f8cb4cc0c984b8a91db6/ipython-5.7.0-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "01f2808ebe78ff2f28dc39be3aa635ca",
-        "sha256": "4e7fb265e0264498bd0d62c6261936a658bf3d38beb8a7b10cd2c6327c62ac2a"
-      },
-      "downloads": -1,
-      "filename": "ipython-5.7.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "01f2808ebe78ff2f28dc39be3aa635ca",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 4977745,
-      "upload_time": "2018-05-10T18:56:17",
-      "upload_time_iso_8601": "2018-05-10T18:56:17.132984Z",
-      "url": "https://files.pythonhosted.org/packages/3c/fd/559fead731a29eaa55cc235c8029807b2520976a937c30e9ee603f3bb566/ipython-5.7.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
