--- tests/repositories/fixtures/pypi.org/json/ipython/7.5.0.json
@@ -1,176 +1,0 @@
-{
-  "info": {
-    "author": "The IPython Development Team",
-    "author_email": "ipython-dev@python.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Framework :: IPython",
-      "Intended Audience :: Developers",
-      "Intended Audience :: Science/Research",
-      "License :: OSI Approved :: BSD License",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3 :: Only",
-      "Topic :: System :: Shells"
-    ],
-    "description": "",
-    "description_content_type": "",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://ipython.org",
-    "keywords": "Interactive,Interpreter,Shell,Embedding",
-    "license": "BSD",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "ipython",
-    "package_url": "https://pypi.org/project/ipython/",
-    "platform": "Linux",
-    "project_url": "https://pypi.org/project/ipython/",
-    "project_urls": {
-      "Documentation": "https://ipython.readthedocs.io/",
-      "Funding": "https://numfocus.org/",
-      "Homepage": "https://ipython.org",
-      "Source": "https://github.com/ipython/ipython",
-      "Tracker": "https://github.com/ipython/ipython/issues"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/ipython/7.5.0/",
-    "requires_dist": [
-      "setuptools (>=18.5)",
-      "jedi (>=0.10)",
-      "decorator",
-      "pickleshare",
-      "traitlets (>=4.2)",
-      "prompt-toolkit (<2.1.0,>=2.0.0)",
-      "pygments",
-      "backcall",
-      "typing; python_version == \"3.4\"",
-      "pexpect; sys_platform != \"win32\"",
-      "appnope; sys_platform == \"darwin\"",
-      "colorama; sys_platform == \"win32\"",
-      "win-unicode-console (>=0.5); sys_platform == \"win32\" and python_version < \"3.6\"",
-      "nbconvert; extra == 'all'",
-      "ipywidgets; extra == 'all'",
-      "pygments; extra == 'all'",
-      "ipykernel; extra == 'all'",
-      "notebook; extra == 'all'",
-      "ipyparallel; extra == 'all'",
-      "requests; extra == 'all'",
-      "Sphinx (>=1.3); extra == 'all'",
-      "nbformat; extra == 'all'",
-      "nose (>=0.10.1); extra == 'all'",
-      "numpy; extra == 'all'",
-      "testpath; extra == 'all'",
-      "qtconsole; extra == 'all'",
-      "Sphinx (>=1.3); extra == 'doc'",
-      "ipykernel; extra == 'kernel'",
-      "nbconvert; extra == 'nbconvert'",
-      "nbformat; extra == 'nbformat'",
-      "notebook; extra == 'notebook'",
-      "ipywidgets; extra == 'notebook'",
-      "ipyparallel; extra == 'parallel'",
-      "qtconsole; extra == 'qtconsole'",
-      "nose (>=0.10.1); extra == 'test'",
-      "requests; extra == 'test'",
-      "testpath; extra == 'test'",
-      "pygments; extra == 'test'",
-      "nbformat; extra == 'test'",
-      "ipykernel; extra == 'test'",
-      "numpy; extra == 'test'"
-    ],
-    "requires_python": ">=3.5",
-    "summary": "IPython: Productive Interactive Computing",
-    "version": "7.5.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "jupyter",
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "bgranger"
-      },
-      {
-        "role": "Owner",
-        "user": "fperez"
-      },
-      {
-        "role": "Owner",
-        "user": "ivanov"
-      },
-      {
-        "role": "Owner",
-        "user": "minrk"
-      },
-      {
-        "role": "Owner",
-        "user": "takowl"
-      },
-      {
-        "role": "Maintainer",
-        "user": "krassowski"
-      },
-      {
-        "role": "Maintainer",
-        "user": "mbussonn"
-      },
-      {
-        "role": "Maintainer",
-        "user": "trallard"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "f40ea889fb7adf989760c5e7a38bd112",
-        "sha256": "1b4c76bf1e8dd9067a4f5ab4695d4c5ad81c30d7d06f7592f4c069c389e37f37"
-      },
-      "downloads": -1,
-      "filename": "ipython-7.5.0-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "f40ea889fb7adf989760c5e7a38bd112",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": ">=3.5",
-      "size": 770001,
-      "upload_time": "2019-04-25T15:38:21",
-      "upload_time_iso_8601": "2019-04-25T15:38:21.726776Z",
-      "url": "https://files.pythonhosted.org/packages/a9/2e/41dce4ed129057e05a555a7f9629aa2d5f81fdcd4d16568bc24b75a1d2c9/ipython-7.5.0-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "0e8c1d7c14f309f6cd2dfd4e48e75cb1",
-        "sha256": "cd2a17ac273fea8bf8953118a2d83bad94f592f0db3e83fff9129a1842e36dbe"
-      },
-      "downloads": -1,
-      "filename": "ipython-7.5.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "0e8c1d7c14f309f6cd2dfd4e48e75cb1",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=3.5",
-      "size": 5118610,
-      "upload_time": "2019-04-25T15:38:33",
-      "upload_time_iso_8601": "2019-04-25T15:38:33.098872Z",
-      "url": "https://files.pythonhosted.org/packages/75/74/9b0ef91c8e356c907bb12297000951acb804583b54eeaddc342c5bad4d96/ipython-7.5.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
