--- tests/repositories/fixtures/pypi.org/json/isort/4.3.4.json
@@ -1,134 +1,0 @@
-{
-  "info": {
-    "author": "Timothy Crosley",
-    "author_email": "timothy.crosley@gmail.com",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 6 - Mature",
-      "Environment :: Console",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Natural Language :: English",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: Implementation :: CPython",
-      "Programming Language :: Python :: Implementation :: PyPy",
-      "Topic :: Software Development :: Libraries",
-      "Topic :: Utilities"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/timothycrosley/isort",
-    "keywords": "Refactor",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "isort",
-    "package_url": "https://pypi.org/project/isort/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/isort/",
-    "project_urls": {
-      "Homepage": "https://github.com/timothycrosley/isort"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/isort/4.3.4/",
-    "requires_dist": null,
-    "requires_python": "",
-    "summary": "A Python utility / library to sort Python imports.",
-    "version": "4.3.4",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "static"
-      },
-      {
-        "role": "Owner",
-        "user": "timothycrosley"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "42bccda292eca3c91eadf3eb781a224f",
-        "sha256": "383c39c10b5db83e8d150ac5b84d74bda96e3a1b06a30257f022dcbcd21f54b9"
-      },
-      "downloads": -1,
-      "filename": "isort-4.3.4-py2-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "42bccda292eca3c91eadf3eb781a224f",
-      "packagetype": "bdist_wheel",
-      "python_version": "2.7",
-      "requires_python": null,
-      "size": 45393,
-      "upload_time": "2018-02-12T15:06:38",
-      "upload_time_iso_8601": "2018-02-12T15:06:38.441257Z",
-      "url": "https://files.pythonhosted.org/packages/41/d8/a945da414f2adc1d9e2f7d6e7445b27f2be42766879062a2e63616ad4199/isort-4.3.4-py2-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "6c3b582d7782633ec23917b00a97a2fe",
-        "sha256": "5668dce9fb48544c57ed626982e190c8ea99e3a612850453e9c3b193b9fa2edc"
-      },
-      "downloads": -1,
-      "filename": "isort-4.3.4-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "6c3b582d7782633ec23917b00a97a2fe",
-      "packagetype": "bdist_wheel",
-      "python_version": "3.6",
-      "requires_python": null,
-      "size": 45352,
-      "upload_time": "2018-02-12T15:06:20",
-      "upload_time_iso_8601": "2018-02-12T15:06:20.089641Z",
-      "url": "https://files.pythonhosted.org/packages/1f/2c/22eee714d7199ae0464beda6ad5fedec8fee6a2f7ffd1e8f1840928fe318/isort-4.3.4-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "9244631852cf8bd8559f7ab78bf4ec78",
-        "sha256": "234ad07e1e2780c27fa56364eefa734bee991b0d744337ef7e7ce3d5b1b59f39"
-      },
-      "downloads": -1,
-      "filename": "isort-4.3.4.tar.gz",
-      "has_sig": false,
-      "md5_digest": "9244631852cf8bd8559f7ab78bf4ec78",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 56070,
-      "upload_time": "2018-02-12T15:06:16",
-      "upload_time_iso_8601": "2018-02-12T15:06:16.498194Z",
-      "url": "https://files.pythonhosted.org/packages/b1/de/a628d16fdba0d38cafb3d7e34d4830f2c9cb3881384ce5c08c44762e1846/isort-4.3.4.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
