--- tests/repositories/fixtures/pypi.org/json/more-itertools/4.1.0.json
@@ -1,132 +1,0 @@
-{
-  "info": {
-    "author": "Erik Rose",
-    "author_email": "erikrose@grinchcentral.com",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Natural Language :: English",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.2",
-      "Programming Language :: Python :: 3.3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Topic :: Software Development :: Libraries"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": "https://pythonhosted.org/more-itertools/",
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/erikrose/more-itertools",
-    "keywords": "itertools,iterator,iteration,filter,peek,peekable,collate,chunk,chunked",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "more-itertools",
-    "package_url": "https://pypi.org/project/more-itertools/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/more-itertools/",
-    "project_urls": {
-      "Homepage": "https://github.com/erikrose/more-itertools"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/more-itertools/4.1.0/",
-    "requires_dist": [
-      "six (<2.0.0,>=1.0.0)"
-    ],
-    "requires_python": "",
-    "summary": "More routines for operating on iterables, beyond itertools",
-    "version": "4.1.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "erikrose"
-      },
-      {
-        "role": "Maintainer",
-        "user": "bbayles"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "703e1e0922de1f11823da60af1488b7a",
-        "sha256": "0f461c2cd4ec16611396f9ee57f40433de3d59e95475d84c0c829cde02f746cd"
-      },
-      "downloads": -1,
-      "filename": "more_itertools-4.1.0-py2-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "703e1e0922de1f11823da60af1488b7a",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2",
-      "requires_python": null,
-      "size": 47987,
-      "upload_time": "2018-01-21T15:34:19",
-      "upload_time_iso_8601": "2018-01-21T15:34:19.304356Z",
-      "url": "https://files.pythonhosted.org/packages/4a/88/c28e2a2da8f3dc3a391d9c97ad949f2ea0c05198222e7e6af176e5bf9b26/more_itertools-4.1.0-py2-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "ae17a45d13e9dc319794c40fa739c38f",
-        "sha256": "580b6002d1f28feb5bcb8303278d59cf17dfbd19a63a5c2375112dae72c9bf98"
-      },
-      "downloads": -1,
-      "filename": "more_itertools-4.1.0-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "ae17a45d13e9dc319794c40fa739c38f",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": null,
-      "size": 47988,
-      "upload_time": "2018-01-21T15:34:20",
-      "upload_time_iso_8601": "2018-01-21T15:34:20.567853Z",
-      "url": "https://files.pythonhosted.org/packages/7a/46/886917c6a4ce49dd3fff250c01c5abac5390d57992751384fe61befc4877/more_itertools-4.1.0-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "bf351a1050242ce3af7e475a4da1a26b",
-        "sha256": "bab2dc6f4be8f9a4a72177842c5283e2dff57c167439a03e3d8d901e854f0f2e"
-      },
-      "downloads": -1,
-      "filename": "more-itertools-4.1.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "bf351a1050242ce3af7e475a4da1a26b",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 51310,
-      "upload_time": "2018-01-21T15:34:22",
-      "upload_time_iso_8601": "2018-01-21T15:34:22.533243Z",
-      "url": "https://files.pythonhosted.org/packages/db/0b/f5660bf6299ec5b9f17bd36096fa8148a1c843fa77ddfddf9bebac9301f7/more-itertools-4.1.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
