--- tests/repositories/fixtures/pypi.org/json/pastel/0.1.0.json
@@ -1,104 +1,0 @@
-{
-  "info": {
-    "author": "Sébastien Eustace",
-    "author_email": "sebastien@eustace.io",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Intended Audience :: Developers",
-      "Operating System :: OS Independent",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: Implementation :: CPython",
-      "Programming Language :: Python :: Implementation :: PyPy"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "https://github.com/sdispater/pastel/archive/0.1.0.tar.gz",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/sdispater/pastel",
-    "keywords": "",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "pastel",
-    "package_url": "https://pypi.org/project/pastel/",
-    "platform": "UNKNOWN",
-    "project_url": "https://pypi.org/project/pastel/",
-    "project_urls": {
-      "Download": "https://github.com/sdispater/pastel/archive/0.1.0.tar.gz",
-      "Homepage": "https://github.com/sdispater/pastel"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/pastel/0.1.0/",
-    "requires_dist": null,
-    "requires_python": "",
-    "summary": "Bring colors to your terminal.",
-    "version": "0.1.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "SDisPater"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "cf7c53ab0a5d7e7c721425b24b486124",
-        "sha256": "754d192c088e256d52a3f825c3b9e14252d5adc70f53656453f6431e50a70b99"
-      },
-      "downloads": -1,
-      "filename": "pastel-0.1.0-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "cf7c53ab0a5d7e7c721425b24b486124",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": null,
-      "size": 6718,
-      "upload_time": "2017-02-01T23:26:20",
-      "upload_time_iso_8601": "2017-02-01T23:26:20.148511Z",
-      "url": "https://files.pythonhosted.org/packages/9b/7e/7d701686013c0d7dae62e0977467232a6adc2e562c23878eb3cd4f97d02e/pastel-0.1.0-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "43ea5f07660f630da18ae1827f5b4333",
-        "sha256": "22f14474c4120b37c54ac2173b49b0ac1de9283ca714be6eb3ea8b39296285a9"
-      },
-      "downloads": -1,
-      "filename": "pastel-0.1.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "43ea5f07660f630da18ae1827f5b4333",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 4490,
-      "upload_time": "2017-02-01T23:26:21",
-      "upload_time_iso_8601": "2017-02-01T23:26:21.720111Z",
-      "url": "https://files.pythonhosted.org/packages/bd/13/a68f2e448b471e8c49e9b596d569ae167a5135ac672b1dc5f24f62f9c15f/pastel-0.1.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
