--- tests/repositories/fixtures/pypi.org/json/pluggy/0.6.0.json
@@ -1,151 +1,0 @@
-{
-  "info": {
-    "author": "Holger Krekel",
-    "author_email": "holger@merlinux.eu",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 4 - Beta",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Operating System :: MacOS :: MacOS X",
-      "Operating System :: Microsoft :: Windows",
-      "Operating System :: POSIX",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: Implementation :: CPython",
-      "Programming Language :: Python :: Implementation :: PyPy",
-      "Topic :: Software Development :: Libraries",
-      "Topic :: Software Development :: Testing",
-      "Topic :: Utilities"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/pytest-dev/pluggy",
-    "keywords": "",
-    "license": "MIT license",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "pluggy",
-    "package_url": "https://pypi.org/project/pluggy/",
-    "platform": "unix",
-    "project_url": "https://pypi.org/project/pluggy/",
-    "project_urls": {
-      "Homepage": "https://github.com/pytest-dev/pluggy"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/pluggy/0.6.0/",
-    "requires_dist": null,
-    "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-    "summary": "plugin and hook calling mechanisms for python",
-    "version": "0.6.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "The_Compiler"
-      },
-      {
-        "role": "Owner",
-        "user": "flub"
-      },
-      {
-        "role": "Owner",
-        "user": "goodboy"
-      },
-      {
-        "role": "Owner",
-        "user": "hpk"
-      },
-      {
-        "role": "Owner",
-        "user": "nicoddemus"
-      },
-      {
-        "role": "Owner",
-        "user": "ronny"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "095eed084713c9b2a9a01520485e20fb",
-        "sha256": "f5f767d398f18aa177976bf9c4d0c05d96487a7d8f07062251585803aaf56246"
-      },
-      "downloads": -1,
-      "filename": "pluggy-0.6.0-py2-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "095eed084713c9b2a9a01520485e20fb",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 11953,
-      "upload_time": "2018-04-15T17:55:28",
-      "upload_time_iso_8601": "2018-04-15T17:55:28.983278Z",
-      "url": "https://files.pythonhosted.org/packages/82/05/43e3947125a2137cba4746135c75934ceed1863f27e050fc560052104a71/pluggy-0.6.0-py2-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "2b6dc266f54023dfb26726686ee6b227",
-        "sha256": "d34798b80853ab688de1a3ca5b99ba4de91c459c19c76a555dc939979ae67eb0"
-      },
-      "downloads": -1,
-      "filename": "pluggy-0.6.0-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "2b6dc266f54023dfb26726686ee6b227",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 13723,
-      "upload_time": "2018-04-15T17:55:22",
-      "upload_time_iso_8601": "2018-04-15T17:55:22.927924Z",
-      "url": "https://files.pythonhosted.org/packages/ba/65/ded3bc40bbf8d887f262f150fbe1ae6637765b5c9534bd55690ed2c0b0f7/pluggy-0.6.0-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "ef8a88abcd501afd47cb22245fe4315a",
-        "sha256": "a982e208d054867661d27c6d2a86b17ba05fbb6b1bdc01f42660732dd107f865"
-      },
-      "downloads": -1,
-      "filename": "pluggy-0.6.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "ef8a88abcd501afd47cb22245fe4315a",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 19678,
-      "upload_time": "2017-11-24T16:33:11",
-      "upload_time_iso_8601": "2017-11-24T16:33:11.250495Z",
-      "url": "https://files.pythonhosted.org/packages/11/bf/cbeb8cdfaffa9f2ea154a30ae31a9d04a1209312e2919138b4171a1f8199/pluggy-0.6.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
