--- tests/repositories/fixtures/pypi.org/json/py/1.5.3.json
@@ -1,132 +1,0 @@
-{
-  "info": {
-    "author": "holger krekel, Ronny Pfannschmidt, Benjamin Peterson and others",
-    "author_email": "pytest-dev@python.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 6 - Mature",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Operating System :: MacOS :: MacOS X",
-      "Operating System :: Microsoft :: Windows",
-      "Operating System :: POSIX",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: Implementation :: CPython",
-      "Programming Language :: Python :: Implementation :: PyPy",
-      "Topic :: Software Development :: Libraries",
-      "Topic :: Software Development :: Testing",
-      "Topic :: Utilities"
-    ],
-    "description": "",
-    "description_content_type": "",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://py.readthedocs.io/",
-    "keywords": "",
-    "license": "MIT license",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "py",
-    "package_url": "https://pypi.org/project/py/",
-    "platform": "unix",
-    "project_url": "https://pypi.org/project/py/",
-    "project_urls": {
-      "Homepage": "http://py.readthedocs.io/"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/py/1.5.3/",
-    "requires_dist": null,
-    "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-    "summary": "library with cross-python path, ini-parsing, io, code, log facilities",
-    "version": "1.5.3",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "pytest-dev",
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "The_Compiler"
-      },
-      {
-        "role": "Owner",
-        "user": "anatoly"
-      },
-      {
-        "role": "Owner",
-        "user": "flub"
-      },
-      {
-        "role": "Owner",
-        "user": "hpk"
-      },
-      {
-        "role": "Owner",
-        "user": "pfctdayelise"
-      },
-      {
-        "role": "Owner",
-        "user": "ronny"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "b316b380701661cb67732ecdaef30eeb",
-        "sha256": "ef4a94f47156178e42ef8f2b131db420e0f4b6aa0b3936b6dbde6ad6487476a5"
-      },
-      "downloads": -1,
-      "filename": "py-1.5.3-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "b316b380701661cb67732ecdaef30eeb",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 84903,
-      "upload_time": "2018-03-22T10:06:50",
-      "upload_time_iso_8601": "2018-03-22T10:06:50.318783Z",
-      "url": "https://files.pythonhosted.org/packages/67/a5/f77982214dd4c8fd104b066f249adea2c49e25e8703d284382eb5e9ab35a/py-1.5.3-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "623e80cfc06df930414a9ce4bf0fd6c9",
-        "sha256": "2df2c513c3af11de15f58189ba5539ddc4768c6f33816dc5c03950c8bd6180fa"
-      },
-      "downloads": -1,
-      "filename": "py-1.5.3.tar.gz",
-      "has_sig": false,
-      "md5_digest": "623e80cfc06df930414a9ce4bf0fd6c9",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 202335,
-      "upload_time": "2018-03-22T10:06:52",
-      "upload_time_iso_8601": "2018-03-22T10:06:52.627078Z",
-      "url": "https://files.pythonhosted.org/packages/f7/84/b4c6e84672c4ceb94f727f3da8344037b62cee960d80e999b1cd9b832d83/py-1.5.3.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
