--- tests/repositories/fixtures/pypi.org/json/pytest/3.5.0.json
@@ -1,143 +1,0 @@
-{
-  "info": {
-    "author": "Holger Krekel, Bruno Oliveira, Ronny Pfannschmidt, Floris Bruynooghe, Brianna Laugher, Florian Bruhin and others",
-    "author_email": "",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 6 - Mature",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Operating System :: MacOS :: MacOS X",
-      "Operating System :: Microsoft :: Windows",
-      "Operating System :: POSIX",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: 3.7",
-      "Topic :: Software Development :: Libraries",
-      "Topic :: Software Development :: Testing",
-      "Topic :: Utilities"
-    ],
-    "description": "",
-    "description_content_type": "",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://pytest.org",
-    "keywords": "test unittest",
-    "license": "MIT license",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "pytest",
-    "package_url": "https://pypi.org/project/pytest/",
-    "platform": "unix",
-    "project_url": "https://pypi.org/project/pytest/",
-    "project_urls": {
-      "Homepage": "http://pytest.org"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/pytest/3.5.0/",
-    "requires_dist": [
-      "py (>=1.5.0)",
-      "six (>=1.10.0)",
-      "setuptools",
-      "attrs (>=17.4.0)",
-      "more-itertools (>=4.0.0)",
-      "pluggy (<0.7,>=0.5)",
-      "funcsigs; python_version < \"3.0\"",
-      "colorama; sys_platform == \"win32\""
-    ],
-    "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-    "summary": "pytest: simple powerful testing with Python",
-    "version": "3.5.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "The_Compiler"
-      },
-      {
-        "role": "Owner",
-        "user": "anatoly"
-      },
-      {
-        "role": "Owner",
-        "user": "flub"
-      },
-      {
-        "role": "Owner",
-        "user": "hpk"
-      },
-      {
-        "role": "Owner",
-        "user": "nicoddemus"
-      },
-      {
-        "role": "Owner",
-        "user": "pfctdayelise"
-      },
-      {
-        "role": "Owner",
-        "user": "ronny"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "4a8651dec151e76f283bf59e333286f9",
-        "sha256": "427b4582bda18e92ad1967e8b1e071e2c53e6cb7e3e5f090cb3ca443455be23f"
-      },
-      "downloads": -1,
-      "filename": "pytest-3.5.0-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "4a8651dec151e76f283bf59e333286f9",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 194247,
-      "upload_time": "2018-03-22T23:47:54",
-      "upload_time_iso_8601": "2018-03-22T23:47:54.595523Z",
-      "url": "https://files.pythonhosted.org/packages/ed/96/271c93f75212c06e2a7ec3e2fa8a9c90acee0a4838dc05bf379ea09aae31/pytest-3.5.0-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "ccd78dac54112045f561c4df86631f19",
-        "sha256": "677b1d6decd29c041fe64276f29f79fbe66e40c59e445eb251366b4a8ab8bf68"
-      },
-      "downloads": -1,
-      "filename": "pytest-3.5.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "ccd78dac54112045f561c4df86631f19",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 830816,
-      "upload_time": "2018-03-22T23:47:56",
-      "upload_time_iso_8601": "2018-03-22T23:47:56.511852Z",
-      "url": "https://files.pythonhosted.org/packages/2d/56/6019153cdd743300c5688ab3b07702355283e53c83fbf922242c053ffb7b/pytest-3.5.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
