--- tests/repositories/fixtures/pypi.org/json/pytest/3.5.1.json
@@ -1,145 +1,0 @@
-{
-  "info": {
-    "author": "Holger Krekel, Bruno Oliveira, Ronny Pfannschmidt, Floris Bruynooghe, Brianna Laugher, Florian Bruhin and others",
-    "author_email": "",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 6 - Mature",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Operating System :: MacOS :: MacOS X",
-      "Operating System :: Microsoft :: Windows",
-      "Operating System :: POSIX",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: 3.7",
-      "Topic :: Software Development :: Libraries",
-      "Topic :: Software Development :: Testing",
-      "Topic :: Utilities"
-    ],
-    "description": "",
-    "description_content_type": "",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://pytest.org",
-    "keywords": "test unittest",
-    "license": "MIT license",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "pytest",
-    "package_url": "https://pypi.org/project/pytest/",
-    "platform": "unix",
-    "project_url": "https://pypi.org/project/pytest/",
-    "project_urls": {
-      "Homepage": "http://pytest.org",
-      "Source": "https://github.com/pytest-dev/pytest",
-      "Tracker": "https://github.com/pytest-dev/pytest/issues"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/pytest/3.5.1/",
-    "requires_dist": [
-      "py (>=1.5.0)",
-      "six (>=1.10.0)",
-      "setuptools",
-      "attrs (>=17.4.0)",
-      "more-itertools (>=4.0.0)",
-      "pluggy (<0.7,>=0.5)",
-      "funcsigs; python_version < \"3.0\"",
-      "colorama; sys_platform == \"win32\""
-    ],
-    "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-    "summary": "pytest: simple powerful testing with Python",
-    "version": "3.5.1",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "The_Compiler"
-      },
-      {
-        "role": "Owner",
-        "user": "anatoly"
-      },
-      {
-        "role": "Owner",
-        "user": "flub"
-      },
-      {
-        "role": "Owner",
-        "user": "hpk"
-      },
-      {
-        "role": "Owner",
-        "user": "nicoddemus"
-      },
-      {
-        "role": "Owner",
-        "user": "pfctdayelise"
-      },
-      {
-        "role": "Owner",
-        "user": "ronny"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "1e81fba94885bef80170545d045924eb",
-        "sha256": "d327df3686046c5b374a9776d9e11606f7dba6fb3db5cf5d60ebc78a31e0768e"
-      },
-      "downloads": -1,
-      "filename": "pytest-3.5.1-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "1e81fba94885bef80170545d045924eb",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 192143,
-      "upload_time": "2018-04-24T21:37:43",
-      "upload_time_iso_8601": "2018-04-24T21:37:43.104462Z",
-      "url": "https://files.pythonhosted.org/packages/76/52/fc48d02492d9e6070cb672d9133382e83084f567f88eff1c27bd2c6c27a8/pytest-3.5.1-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "961104636090457187851ccb9ef0f677",
-        "sha256": "b8fe151f3e181801dd38583a1c03818fbc662a8fce96c9063a0af624613e78f8"
-      },
-      "downloads": -1,
-      "filename": "pytest-3.5.1.tar.gz",
-      "has_sig": false,
-      "md5_digest": "961104636090457187851ccb9ef0f677",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 830571,
-      "upload_time": "2018-04-24T21:37:44",
-      "upload_time_iso_8601": "2018-04-24T21:37:44.492084Z",
-      "url": "https://files.pythonhosted.org/packages/b2/85/24954df0ea8156599563b753de54383a5d702081093b7953334e4701b8d8/pytest-3.5.1.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
