--- tests/repositories/fixtures/pypi.org/json/requests/2.18.0.json
@@ -1,128 +1,0 @@
-{
-  "info": {
-    "author": "Kenneth Reitz",
-    "author_email": "me@kennethreitz.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: Apache Software License",
-      "Natural Language :: English",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2.6",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: Implementation :: CPython",
-      "Programming Language :: Python :: Implementation :: PyPy"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://python-requests.org",
-    "keywords": "",
-    "license": "Apache 2.0",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "requests",
-    "package_url": "https://pypi.org/project/requests/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/requests/",
-    "project_urls": {
-      "Homepage": "http://python-requests.org"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/requests/2.18.0/",
-    "requires_dist": [
-      "certifi (>=2017.4.17)",
-      "chardet (>=3.0.2,<3.1.0)",
-      "idna (>=2.5,<2.6)",
-      "urllib3 (<1.22,>=1.21.1)",
-      "cryptography (>=1.3.4); extra == 'security'",
-      "idna (>=2.0.0); extra == 'security'",
-      "pyOpenSSL (>=0.14); extra == 'security'",
-      "PySocks (!=1.5.7,>=1.5.6); extra == 'socks'",
-      "win-inet-pton; sys_platform == \"win32\" and (python_version == \"2.7\" or python_version == \"2.6\") and extra == 'socks'"
-    ],
-    "requires_python": "",
-    "summary": "Python HTTP for Humans.",
-    "version": "2.18.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "Lukasa"
-      },
-      {
-        "role": "Owner",
-        "user": "graffatcolmingov"
-      },
-      {
-        "role": "Owner",
-        "user": "nateprewitt"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "e2f0c81405acbf53d0412b984eb3fc578cdd10e347374e1aec074638a500c186",
-        "md5": "6f34e2439fcb3dd1b6e3304903bb6be8",
-        "sha256": "5e88d64aa56ac0fda54e77fb9762ebc65879e171b746d5479a33c4082519d6c6"
-      },
-      "downloads": -1,
-      "filename": "requests-2.18.0-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "6f34e2439fcb3dd1b6e3304903bb6be8",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": null,
-      "size": 563596,
-      "upload_time": "2017-06-14T15:44:35",
-      "upload_time_iso_8601": "2017-06-14T15:44:35.080617Z",
-      "url": "https://files.pythonhosted.org/packages/e2/f0/c81405acbf53d0412b984eb3fc578cdd10e347374e1aec074638a500c186/requests-2.18.0-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "e097e2f972b6826c9cfe57b6934e3773d2783733bc2d345d810bafd309df3d15",
-        "md5": "b8b333ace1653652ddcce95284577f5c",
-        "sha256": "cd0189f962787284bff715fddaad478eb4d9c15aa167bd64e52ea0f661e7ea5c"
-      },
-      "downloads": -1,
-      "filename": "requests-2.18.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "b8b333ace1653652ddcce95284577f5c",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 124085,
-      "upload_time": "2017-06-14T15:44:37",
-      "upload_time_iso_8601": "2017-06-14T15:44:37.484470Z",
-      "url": "https://files.pythonhosted.org/packages/e0/97/e2f972b6826c9cfe57b6934e3773d2783733bc2d345d810bafd309df3d15/requests-2.18.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
