--- tests/repositories/fixtures/pypi.org/json/requests/2.18.1.json
@@ -1,128 +1,0 @@
-{
-  "info": {
-    "author": "Kenneth Reitz",
-    "author_email": "me@kennethreitz.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: Apache Software License",
-      "Natural Language :: English",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2.6",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: Implementation :: CPython",
-      "Programming Language :: Python :: Implementation :: PyPy"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://python-requests.org",
-    "keywords": "",
-    "license": "Apache 2.0",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "requests",
-    "package_url": "https://pypi.org/project/requests/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/requests/",
-    "project_urls": {
-      "Homepage": "http://python-requests.org"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/requests/2.18.1/",
-    "requires_dist": [
-      "certifi (>=2017.4.17)",
-      "chardet (>=3.0.2,<3.1.0)",
-      "idna (>=2.5,<2.6)",
-      "urllib3 (<1.22,>=1.21.1)",
-      "cryptography (>=1.3.4); extra == 'security'",
-      "idna (>=2.0.0); extra == 'security'",
-      "pyOpenSSL (>=0.14); extra == 'security'",
-      "PySocks (!=1.5.7,>=1.5.6); extra == 'socks'",
-      "win-inet-pton; sys_platform == \"win32\" and (python_version == \"2.7\" or python_version == \"2.6\") and extra == 'socks'"
-    ],
-    "requires_python": "",
-    "summary": "Python HTTP for Humans.",
-    "version": "2.18.1",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "Lukasa"
-      },
-      {
-        "role": "Owner",
-        "user": "graffatcolmingov"
-      },
-      {
-        "role": "Owner",
-        "user": "nateprewitt"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "5a58671011e3ff4a06e2969322267d78dcfda1bf4d1576551df1cce93cd7239d",
-        "md5": "a7fbdc82134a2610b3d0cdc7e59f0bde",
-        "sha256": "6afd3371c1f4c1970497cdcace5c5ecbbe58267bf05ca1abd93d99d170803ab7"
-      },
-      "downloads": -1,
-      "filename": "requests-2.18.1-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "a7fbdc82134a2610b3d0cdc7e59f0bde",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": null,
-      "size": 88107,
-      "upload_time": "2017-06-14T17:51:25",
-      "upload_time_iso_8601": "2017-06-14T17:51:25.096686Z",
-      "url": "https://files.pythonhosted.org/packages/5a/58/671011e3ff4a06e2969322267d78dcfda1bf4d1576551df1cce93cd7239d/requests-2.18.1-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "2cb52b6e8ef8dd18203b6399e9f28c7d54f6de7b7549853fe36d575bd31e29a7",
-        "md5": "40f723ed01dddeaf990d0609d073f021",
-        "sha256": "c6f3bdf4a4323ac7b45d01e04a6f6c20e32a052cd04de81e05103abc049ad9b9"
-      },
-      "downloads": -1,
-      "filename": "requests-2.18.1.tar.gz",
-      "has_sig": false,
-      "md5_digest": "40f723ed01dddeaf990d0609d073f021",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 124229,
-      "upload_time": "2017-06-14T17:51:28",
-      "upload_time_iso_8601": "2017-06-14T17:51:28.960131Z",
-      "url": "https://files.pythonhosted.org/packages/2c/b5/2b6e8ef8dd18203b6399e9f28c7d54f6de7b7549853fe36d575bd31e29a7/requests-2.18.1.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
