--- tests/repositories/fixtures/pypi.org/json/requests/2.18.4.json
@@ -1,125 +1,0 @@
-{
-  "info": {
-    "author": "Kenneth Reitz",
-    "author_email": "me@kennethreitz.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: Apache Software License",
-      "Natural Language :: English",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2.6",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: Implementation :: CPython",
-      "Programming Language :: Python :: Implementation :: PyPy"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://python-requests.org",
-    "keywords": "",
-    "license": "Apache 2.0",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "requests",
-    "package_url": "https://pypi.org/project/requests/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/requests/",
-    "project_urls": {
-      "Homepage": "http://python-requests.org"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/requests/2.18.4/",
-    "requires_dist": [
-      "certifi (>=2017.4.17)",
-      "chardet (>=3.0.2,<3.1.0)",
-      "idna (>=2.5,<2.7)",
-      "urllib3 (<1.23,>=1.21.1)",
-      "cryptography (>=1.3.4); extra == 'security'",
-      "idna (>=2.0.0); extra == 'security'",
-      "pyOpenSSL (>=0.14); extra == 'security'",
-      "PySocks (!=1.5.7,>=1.5.6); extra == 'socks'",
-      "win-inet-pton; sys_platform == \"win32\" and (python_version == \"2.7\" or python_version == \"2.6\") and extra == 'socks'"
-    ],
-    "requires_python": "",
-    "summary": "Python HTTP for Humans.",
-    "version": "2.18.4",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "Lukasa"
-      },
-      {
-        "role": "Owner",
-        "user": "graffatcolmingov"
-      },
-      {
-        "role": "Owner",
-        "user": "nateprewitt"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "e770e65750c42f40b97b0ed738d0f859",
-        "sha256": "098be851f30be5bcb2c7537798d44314f576e53818ba9def25141ae4dce8b25d"
-      },
-      "downloads": -1,
-      "filename": "requests-2.18.4-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "e770e65750c42f40b97b0ed738d0f859",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": null,
-      "size": 88704,
-      "upload_time": "2017-08-15T13:23:43",
-      "upload_time_iso_8601": "2017-08-15T13:23:43.489631Z",
-      "url": "https://files.pythonhosted.org/packages/49/df/50aa1999ab9bde74656c2919d9c0c085fd2b3775fd3eca826012bef76d8c/requests-2.18.4-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "942a6a383dc94da90cf58f5adcf028a4",
-        "sha256": "ec62f7e0e9d4814656b0172dbd592fea06127c6556ff5651eb5d2c8768671fd4"
-      },
-      "downloads": -1,
-      "filename": "requests-2.18.4.tar.gz",
-      "has_sig": false,
-      "md5_digest": "942a6a383dc94da90cf58f5adcf028a4",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 126224,
-      "upload_time": "2017-08-15T13:23:46",
-      "upload_time_iso_8601": "2017-08-15T13:23:46.348325Z",
-      "url": "https://files.pythonhosted.org/packages/b0/e1/eab4fc3752e3d240468a8c0b284607899d2fbfb236a56b7377a329aa8d09/requests-2.18.4.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
