--- tests/repositories/fixtures/pypi.org/json/requests/2.19.0.json
@@ -1,127 +1,0 @@
-{
-  "info": {
-    "author": "Kenneth Reitz",
-    "author_email": "me@kennethreitz.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: Apache Software License",
-      "Natural Language :: English",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: Implementation :: CPython",
-      "Programming Language :: Python :: Implementation :: PyPy"
-    ],
-    "description": "",
-    "description_content_type": "",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://python-requests.org",
-    "keywords": "",
-    "license": "Apache 2.0",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "requests",
-    "package_url": "https://pypi.org/project/requests/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/requests/",
-    "project_urls": {
-      "Homepage": "http://python-requests.org"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/requests/2.19.0/",
-    "requires_dist": [
-      "chardet (<3.1.0,>=3.0.2)",
-      "idna (<2.8,>=2.5)",
-      "urllib3 (<1.24,>=1.21.1)",
-      "certifi (>=2017.4.17)",
-      "pyOpenSSL (>=0.14); extra == 'security'",
-      "cryptography (>=1.3.4); extra == 'security'",
-      "idna (>=2.0.0); extra == 'security'",
-      "PySocks (!=1.5.7,>=1.5.6); extra == 'socks'",
-      "win-inet-pton; sys_platform == \"win32\" and (python_version == \"2.7\" or python_version == \"2.6\") and extra == 'socks'"
-    ],
-    "requires_python": ">=2.6, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-    "summary": "Python HTTP for Humans.",
-    "version": "2.19.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "Lukasa"
-      },
-      {
-        "role": "Owner",
-        "user": "graffatcolmingov"
-      },
-      {
-        "role": "Owner",
-        "user": "nateprewitt"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "cc15e1c318dbc20032ffbe5628837ca0de2d5b116ffd1b849c699634010f6a5d",
-        "md5": "1ae2b89f8ea3e4aea8b199987fb2aae9",
-        "sha256": "421cfc8d9dde7d6aff68196420afd86b88c65d77d8da9cf83f4ecad785d7b9d6"
-      },
-      "downloads": -1,
-      "filename": "requests-2.19.0-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "1ae2b89f8ea3e4aea8b199987fb2aae9",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": ">=2.6, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 91865,
-      "upload_time": "2018-06-12T14:46:15",
-      "upload_time_iso_8601": "2018-06-12T14:46:15.289074Z",
-      "url": "https://files.pythonhosted.org/packages/cc/15/e1c318dbc20032ffbe5628837ca0de2d5b116ffd1b849c699634010f6a5d/requests-2.19.0-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "752782da3fa4ea7a8c3526c48eaafe427352ff9c931633b917c2251826a43697",
-        "md5": "8a7844c58d496e9e92481de459830229",
-        "sha256": "cc408268d0e21589bcc2b2c248e42932b8c4d112f499c12c92e99e2178a6134c"
-      },
-      "downloads": -1,
-      "filename": "requests-2.19.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "8a7844c58d496e9e92481de459830229",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=2.6, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 130875,
-      "upload_time": "2018-06-12T14:46:17",
-      "upload_time_iso_8601": "2018-06-12T14:46:17.223245Z",
-      "url": "https://files.pythonhosted.org/packages/75/27/82da3fa4ea7a8c3526c48eaafe427352ff9c931633b917c2251826a43697/requests-2.19.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
