--- tests/repositories/fixtures/pypi.org/json/setuptools/39.2.0.json
@@ -1,123 +1,0 @@
-{
-  "info": {
-    "author": "Python Packaging Authority",
-    "author_email": "distutils-sig@python.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Operating System :: OS Independent",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Topic :: Software Development :: Libraries :: Python Modules",
-      "Topic :: System :: Archiving :: Packaging",
-      "Topic :: System :: Systems Administration",
-      "Topic :: Utilities"
-    ],
-    "description": "",
-    "description_content_type": "text/x-rst; charset=UTF-8",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/pypa/setuptools",
-    "keywords": "CPAN PyPI distutils eggs package management",
-    "license": "",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "setuptools",
-    "package_url": "https://pypi.org/project/setuptools/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/setuptools/",
-    "project_urls": {
-      "Documentation": "https://setuptools.readthedocs.io/",
-      "Homepage": "https://github.com/pypa/setuptools"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/setuptools/39.2.0/",
-    "requires_dist": [
-      "certifi (==2016.9.26); extra == 'certs'",
-      "wincertstore (==0.2); (sys_platform=='win32') and extra == 'ssl'"
-    ],
-    "requires_python": ">=2.7,!=3.0.*,!=3.1.*,!=3.2.*",
-    "summary": "Easily download, build, install, upgrade, and uninstall Python packages",
-    "version": "39.2.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "jaraco"
-      },
-      {
-        "role": "Maintainer",
-        "user": "abravalheri"
-      },
-      {
-        "role": "Maintainer",
-        "user": "dstufft"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "7fe1820d941153923aac1d49d7fc37e17b6e73bfbd2904959fffbad77900cf92",
-        "md5": "8d066d2201311ed30be535b473e32fed",
-        "sha256": "8fca9275c89964f13da985c3656cb00ba029d7f3916b37990927ffdf264e7926"
-      },
-      "downloads": -1,
-      "filename": "setuptools-39.2.0-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "8d066d2201311ed30be535b473e32fed",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": ">=2.7,!=3.0.*,!=3.1.*,!=3.2.*",
-      "size": 567556,
-      "upload_time": "2018-05-19T19:19:22",
-      "upload_time_iso_8601": "2018-05-19T19:19:22.625819Z",
-      "url": "https://files.pythonhosted.org/packages/7f/e1/820d941153923aac1d49d7fc37e17b6e73bfbd2904959fffbad77900cf92/setuptools-39.2.0-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "1a04d6f1159feaccdfc508517dba1929eb93a2854de729fa68da9d5c6b48fa00",
-        "md5": "dd4e3fa83a21bf7bf9c51026dc8a4e59",
-        "sha256": "f7cddbb5f5c640311eb00eab6e849f7701fa70bf6a183fc8a2c33dd1d1672fb2"
-      },
-      "downloads": -1,
-      "filename": "setuptools-39.2.0.zip",
-      "has_sig": false,
-      "md5_digest": "dd4e3fa83a21bf7bf9c51026dc8a4e59",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=2.7,!=3.0.*,!=3.1.*,!=3.2.*",
-      "size": 851112,
-      "upload_time": "2018-05-19T19:19:24",
-      "upload_time_iso_8601": "2018-05-19T19:19:24.480740Z",
-      "url": "https://files.pythonhosted.org/packages/1a/04/d6f1159feaccdfc508517dba1929eb93a2854de729fa68da9d5c6b48fa00/setuptools-39.2.0.zip",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
