--- tests/repositories/fixtures/pypi.org/json/setuptools/67.6.1.json
@@ -1,159 +1,0 @@
-{
-  "info": {
-    "author": "Python Packaging Authority",
-    "author_email": "distutils-sig@python.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3 :: Only",
-      "Topic :: Software Development :: Libraries :: Python Modules",
-      "Topic :: System :: Archiving :: Packaging",
-      "Topic :: System :: Systems Administration",
-      "Topic :: Utilities"
-    ],
-    "description": "",
-    "description_content_type": "",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/pypa/setuptools",
-    "keywords": "CPAN PyPI distutils eggs package management",
-    "license": "",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "setuptools",
-    "package_url": "https://pypi.org/project/setuptools/",
-    "platform": null,
-    "project_url": "https://pypi.org/project/setuptools/",
-    "project_urls": {
-      "Changelog": "https://setuptools.pypa.io/en/stable/history.html",
-      "Documentation": "https://setuptools.pypa.io/",
-      "Homepage": "https://github.com/pypa/setuptools"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/setuptools/67.6.1/",
-    "requires_dist": [
-      "sphinx (>=3.5) ; extra == 'docs'",
-      "jaraco.packaging (>=9) ; extra == 'docs'",
-      "rst.linker (>=1.9) ; extra == 'docs'",
-      "furo ; extra == 'docs'",
-      "sphinx-lint ; extra == 'docs'",
-      "jaraco.tidelift (>=1.4) ; extra == 'docs'",
-      "pygments-github-lexers (==0.0.5) ; extra == 'docs'",
-      "sphinx-favicon ; extra == 'docs'",
-      "sphinx-inline-tabs ; extra == 'docs'",
-      "sphinx-reredirects ; extra == 'docs'",
-      "sphinxcontrib-towncrier ; extra == 'docs'",
-      "sphinx-notfound-page (==0.8.3) ; extra == 'docs'",
-      "sphinx-hoverxref (<2) ; extra == 'docs'",
-      "pytest (>=6) ; extra == 'testing'",
-      "pytest-checkdocs (>=2.4) ; extra == 'testing'",
-      "flake8 (<5) ; extra == 'testing'",
-      "pytest-enabler (>=1.3) ; extra == 'testing'",
-      "pytest-perf ; extra == 'testing'",
-      "flake8-2020 ; extra == 'testing'",
-      "virtualenv (>=13.0.0) ; extra == 'testing'",
-      "wheel ; extra == 'testing'",
-      "pip (>=19.1) ; extra == 'testing'",
-      "jaraco.envs (>=2.2) ; extra == 'testing'",
-      "pytest-xdist ; extra == 'testing'",
-      "jaraco.path (>=3.2.0) ; extra == 'testing'",
-      "build[virtualenv] ; extra == 'testing'",
-      "filelock (>=3.4.0) ; extra == 'testing'",
-      "pip-run (>=8.8) ; extra == 'testing'",
-      "ini2toml[lite] (>=0.9) ; extra == 'testing'",
-      "tomli-w (>=1.0.0) ; extra == 'testing'",
-      "pytest-timeout ; extra == 'testing'",
-      "pytest ; extra == 'testing-integration'",
-      "pytest-xdist ; extra == 'testing-integration'",
-      "pytest-enabler ; extra == 'testing-integration'",
-      "virtualenv (>=13.0.0) ; extra == 'testing-integration'",
-      "tomli ; extra == 'testing-integration'",
-      "wheel ; extra == 'testing-integration'",
-      "jaraco.path (>=3.2.0) ; extra == 'testing-integration'",
-      "jaraco.envs (>=2.2) ; extra == 'testing-integration'",
-      "build[virtualenv] ; extra == 'testing-integration'",
-      "filelock (>=3.4.0) ; extra == 'testing-integration'",
-      "pytest-black (>=0.3.7) ; (platform_python_implementation != \"PyPy\") and extra == 'testing'",
-      "pytest-cov ; (platform_python_implementation != \"PyPy\") and extra == 'testing'",
-      "pytest-mypy (>=0.9.1) ; (platform_python_implementation != \"PyPy\") and extra == 'testing'",
-      "pytest-flake8 ; (python_version < \"3.12\") and extra == 'testing'"
-    ],
-    "requires_python": ">=3.7",
-    "summary": "Easily download, build, install, upgrade, and uninstall Python packages",
-    "version": "67.6.1",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "jaraco"
-      },
-      {
-        "role": "Maintainer",
-        "user": "abravalheri"
-      },
-      {
-        "role": "Maintainer",
-        "user": "dstufft"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "3b5b846e000da033d54eeaaf7915126e",
-        "sha256": "e728ca814a823bf7bf60162daf9db95b93d532948c4c0bea762ce62f60189078"
-      },
-      "downloads": -1,
-      "filename": "setuptools-67.6.1-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "3b5b846e000da033d54eeaaf7915126e",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": ">=3.7",
-      "size": 1089263,
-      "upload_time": "2023-03-28T13:45:43",
-      "upload_time_iso_8601": "2023-03-28T13:45:43.525946Z",
-      "url": "https://files.pythonhosted.org/packages/0b/fc/8781442def77b0aa22f63f266d4dadd486ebc0c5371d6290caf4320da4b7/setuptools-67.6.1-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "ee2562f783544d1f95022c906dd3cf98",
-        "sha256": "a737d365c957dd3fced9ddd246118e95dce7a62c3dc49f37e7fdd9e93475d785"
-      },
-      "downloads": -1,
-      "filename": "setuptools-67.6.1.tar.gz",
-      "has_sig": false,
-      "md5_digest": "ee2562f783544d1f95022c906dd3cf98",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=3.7",
-      "size": 2486256,
-      "upload_time": "2023-03-28T13:45:45",
-      "upload_time_iso_8601": "2023-03-28T13:45:45.967259Z",
-      "url": "https://files.pythonhosted.org/packages/cb/46/22ec35f286a77e6b94adf81b4f0d59f402ed981d4251df0ba7b992299146/setuptools-67.6.1.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
