--- tests/repositories/fixtures/pypi.org/json/six/1.11.0.json
@@ -1,100 +1,0 @@
-{
-  "info": {
-    "author": "Benjamin Peterson",
-    "author_email": "benjamin@python.org",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 3",
-      "Topic :: Software Development :: Libraries",
-      "Topic :: Utilities"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://pypi.python.org/pypi/six/",
-    "keywords": "",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "six",
-    "package_url": "https://pypi.org/project/six/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/six/",
-    "project_urls": {
-      "Homepage": "http://pypi.python.org/pypi/six/"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/six/1.11.0/",
-    "requires_dist": null,
-    "requires_python": "",
-    "summary": "Python 2 and 3 compatibility utilities",
-    "version": "1.11.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "gutworth"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "9500094701f7201ddd065c60abcefef1",
-        "sha256": "534e9875e44a507adec601c29b3cbd2ca6dae7df92bf3dd20c7289b2f99f7466"
-      },
-      "downloads": -1,
-      "filename": "six-1.11.0-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "9500094701f7201ddd065c60abcefef1",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": null,
-      "size": 10702,
-      "upload_time": "2017-09-17T18:46:53",
-      "upload_time_iso_8601": "2017-09-17T18:46:53.702194Z",
-      "url": "https://files.pythonhosted.org/packages/67/4b/141a581104b1f6397bfa78ac9d43d8ad29a7ca43ea90a2d863fe3056e86a/six-1.11.0-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "25d3568604f921dd23532b88a0ce17e7",
-        "sha256": "268a4ccb159c1a2d2c79336b02e75058387b0cdbb4cea2f07846a758f48a356d"
-      },
-      "downloads": -1,
-      "filename": "six-1.11.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "25d3568604f921dd23532b88a0ce17e7",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 29860,
-      "upload_time": "2017-09-17T18:46:54",
-      "upload_time_iso_8601": "2017-09-17T18:46:54.492027Z",
-      "url": "https://files.pythonhosted.org/packages/16/d8/bc6316cf98419719bd59c91742194c111b6f2e85abac88e496adefaf7afe/six-1.11.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
