--- tests/repositories/fixtures/pypi.org/json/toga/0.3.0.json
@@ -1,114 +1,0 @@
-{
-  "info": {
-    "author": "Russell Keith-Magee",
-    "author_email": "russell@keith-magee.com",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 3 - Alpha",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: BSD License",
-      "Operating System :: OS Independent",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3 :: Only",
-      "Programming Language :: Python :: 3.10",
-      "Programming Language :: Python :: 3.11",
-      "Programming Language :: Python :: 3.12",
-      "Programming Language :: Python :: 3.7",
-      "Programming Language :: Python :: 3.8",
-      "Programming Language :: Python :: 3.9",
-      "Topic :: Software Development",
-      "Topic :: Software Development :: User Interfaces",
-      "Topic :: Software Development :: Widget Sets"
-    ],
-    "description": "",
-    "description_content_type": "text/x-rst; charset=UTF-8",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://beeware.org/project/projects/libraries/toga/",
-    "keywords": "gui,widget,cross-platform,desktop,mobile,web,macOS,cocoa,iOS,android,windows,winforms,linux,gtk",
-    "license": "New BSD",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "BeeWare Team",
-    "maintainer_email": "team@beeware.org",
-    "name": "toga",
-    "package_url": "https://pypi.org/project/toga/",
-    "platform": null,
-    "project_url": "https://pypi.org/project/toga/",
-    "project_urls": {
-      "Documentation": "http://toga.readthedocs.io/en/latest/",
-      "Funding": "https://beeware.org/contributing/membership/",
-      "Homepage": "https://beeware.org/project/projects/libraries/toga/",
-      "Source": "https://github.com/beeware/toga",
-      "Tracker": "https://github.com/beeware/toga/issues"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/toga/0.3.0/",
-    "requires_dist": [
-      "toga-cocoa (==0.3.0) ; sys_platform==\"darwin\"",
-      "toga-gtk (==0.3.0) ; sys_platform==\"linux\"",
-      "toga-winforms (==0.3.0) ; sys_platform==\"win32\""
-    ],
-    "requires_python": ">=3.7",
-    "summary": "A Python native, OS native GUI toolkit.",
-    "version": "0.3.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "beeware",
-    "roles": []
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "401df3f98e3811ef657aa2acabe039fcbbcae7fc3463f3e18a2458e099465a49",
-        "md5": "4c9f80d34e8d3f6a3cb301749fbca4ca",
-        "sha256": "846f1598d5ce91e44aefaac8abc9ba8c5a7c3bbd8c69d7004b92ef64228500ad"
-      },
-      "downloads": -1,
-      "filename": "toga-0.3.0-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "4c9f80d34e8d3f6a3cb301749fbca4ca",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": ">=3.7",
-      "size": 1945,
-      "upload_time": "2023-01-30T03:29:31",
-      "upload_time_iso_8601": "2023-01-30T03:29:31.034500Z",
-      "url": "https://files.pythonhosted.org/packages/40/1d/f3f98e3811ef657aa2acabe039fcbbcae7fc3463f3e18a2458e099465a49/toga-0.3.0-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "f3cf44b2d505e13b2ec0c7cef2677ac5d8bffb1fb22c4c1d3671179f8a52a40b",
-        "md5": "f6d6ddeffa5c08a875d436b2dc835e1d",
-        "sha256": "0d1f04d6b5773b8682e5eb1260ec9f07b43f950b89916f9406b2db33cde27240"
-      },
-      "downloads": -1,
-      "filename": "toga-0.3.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "f6d6ddeffa5c08a875d436b2dc835e1d",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=3.7",
-      "size": 3236,
-      "upload_time": "2023-01-30T03:29:34",
-      "upload_time_iso_8601": "2023-01-30T03:29:34.595566Z",
-      "url": "https://files.pythonhosted.org/packages/f3/cf/44b2d505e13b2ec0c7cef2677ac5d8bffb1fb22c4c1d3671179f8a52a40b/toga-0.3.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
