--- tests/repositories/fixtures/pypi.org/json/toga/0.3.0dev1.json
@@ -1,107 +1,0 @@
-{
-  "info": {
-    "author": "Russell Keith-Magee",
-    "author_email": "russell@keith-magee.com",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 3 - Alpha",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: BSD License",
-      "Operating System :: OS Independent",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3 :: Only",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Topic :: Software Development",
-      "Topic :: Software Development :: User Interfaces",
-      "Topic :: Software Development :: Widget Sets"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://pybee.org/toga",
-    "keywords": "",
-    "license": "New BSD",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "toga",
-    "package_url": "https://pypi.org/project/toga/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/toga/",
-    "project_urls": {
-      "Homepage": "http://pybee.org/toga"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/toga/0.3.0.dev1/",
-    "requires_dist": [
-      "toga-cocoa; sys_platform==\"darwin\"",
-      "toga-gtk; sys_platform==\"linux\"",
-      "toga-winforms; sys_platform==\"win32\""
-    ],
-    "requires_python": "",
-    "summary": "A Python native, OS native GUI toolkit.",
-    "version": "0.3.0.dev1",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "beeware",
-    "roles": []
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "ebefea806c706d3dc90d4bc8c412c0ad3515fd018074f5fdd4bd020bdd4c0c80",
-        "md5": "7b219b2249b825f28051aae0230f2818",
-        "sha256": "8553bf332d8fbf39b500745ed9c4044a846fbba68e31de70e6fe83fdffcb0a9e"
-      },
-      "downloads": -1,
-      "filename": "toga-0.3.0.dev1-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "7b219b2249b825f28051aae0230f2818",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": null,
-      "size": 4789,
-      "upload_time": "2018-01-14T04:10:57",
-      "upload_time_iso_8601": "2018-01-14T04:10:57.825822Z",
-      "url": "https://files.pythonhosted.org/packages/eb/ef/ea806c706d3dc90d4bc8c412c0ad3515fd018074f5fdd4bd020bdd4c0c80/toga-0.3.0.dev1-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "4f71c55c15950f7275e761fe53fb0dc83fe4f5fba6199d7b8fb05d741dd33566",
-        "md5": "21b8fff4d110ddfb8c3eb939d7b35a1e",
-        "sha256": "4e5c77056792168a4e84c84bb7214dfb614b79f289dcbe1525be614483496439"
-      },
-      "downloads": -1,
-      "filename": "toga-0.3.0.dev1.tar.gz",
-      "has_sig": false,
-      "md5_digest": "21b8fff4d110ddfb8c3eb939d7b35a1e",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 40187,
-      "upload_time": "2018-01-14T04:11:03",
-      "upload_time_iso_8601": "2018-01-14T04:11:03.212494Z",
-      "url": "https://files.pythonhosted.org/packages/4f/71/c55c15950f7275e761fe53fb0dc83fe4f5fba6199d7b8fb05d741dd33566/toga-0.3.0.dev1.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
