--- tests/repositories/fixtures/pypi.org/json/toga/0.3.0dev2.json
@@ -1,107 +1,0 @@
-{
-  "info": {
-    "author": "Russell Keith-Magee",
-    "author_email": "russell@keith-magee.com",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 3 - Alpha",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: BSD License",
-      "Operating System :: OS Independent",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3 :: Only",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Topic :: Software Development",
-      "Topic :: Software Development :: User Interfaces",
-      "Topic :: Software Development :: Widget Sets"
-    ],
-    "description": "",
-    "description_content_type": null,
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://pybee.org/toga",
-    "keywords": "",
-    "license": "New BSD",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "toga",
-    "package_url": "https://pypi.org/project/toga/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/toga/",
-    "project_urls": {
-      "Homepage": "http://pybee.org/toga"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/toga/0.3.0.dev2/",
-    "requires_dist": [
-      "toga-cocoa; sys_platform==\"darwin\"",
-      "toga-gtk; sys_platform==\"linux\"",
-      "toga-winforms; sys_platform==\"win32\""
-    ],
-    "requires_python": "",
-    "summary": "A Python native, OS native GUI toolkit.",
-    "version": "0.3.0.dev2",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "beeware",
-    "roles": []
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "3a68d1f6feb2ded26b9f6c36cd2a826e895e0fa6bba5fe489ec30b9f3bc1dbea",
-        "md5": "1aa1d5f48b81475569ea80ea04db8852",
-        "sha256": "6e0a2f800a351bbe8639802954d8d283a52b8cdde378541610ff2bfb3b24ad2f"
-      },
-      "downloads": -1,
-      "filename": "toga-0.3.0.dev2-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "1aa1d5f48b81475569ea80ea04db8852",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": null,
-      "size": 4788,
-      "upload_time": "2018-01-14T04:52:03",
-      "upload_time_iso_8601": "2018-01-14T04:52:03.658804Z",
-      "url": "https://files.pythonhosted.org/packages/3a/68/d1f6feb2ded26b9f6c36cd2a826e895e0fa6bba5fe489ec30b9f3bc1dbea/toga-0.3.0.dev2-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "blake2b_256": "e36a3264b3d48733cac7546fee02fbc516621574252f7d86546255532b095415",
-        "md5": "a6558d0c5ba3cd763084e564001e9d24",
-        "sha256": "630d2f932bf7aba3a143d3a332190a46a0a3895f509099a20f033caadf131b76"
-      },
-      "downloads": -1,
-      "filename": "toga-0.3.0.dev2.tar.gz",
-      "has_sig": false,
-      "md5_digest": "a6558d0c5ba3cd763084e564001e9d24",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 40203,
-      "upload_time": "2018-01-14T04:52:09",
-      "upload_time_iso_8601": "2018-01-14T04:52:09.347038Z",
-      "url": "https://files.pythonhosted.org/packages/e3/6a/3264b3d48733cac7546fee02fbc516621574252f7d86546255532b095415/toga-0.3.0.dev2.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
