--- tests/repositories/fixtures/pypi.org/json/tomlkit/0.5.2.json
@@ -1,107 +1,0 @@
-{
-  "info": {
-    "author": "Sébastien Eustace",
-    "author_email": "sebastien@eustace.io",
-    "bugtrack_url": null,
-    "classifiers": [
-      "License :: OSI Approved :: MIT License",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: 3.7"
-    ],
-    "description": "",
-    "description_content_type": "text/markdown",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/sdispater/tomlkit",
-    "keywords": "",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "Sébastien Eustace",
-    "maintainer_email": "sebastien@eustace.io",
-    "name": "tomlkit",
-    "package_url": "https://pypi.org/project/tomlkit/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/tomlkit/",
-    "project_urls": {
-      "Homepage": "https://github.com/sdispater/tomlkit",
-      "Repository": "https://github.com/sdispater/tomlkit"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/tomlkit/0.5.2/",
-    "requires_dist": [
-      "enum34 (>=1.1,<2.0); python_version >= \"2.7\" and python_version < \"2.8\"",
-      "functools32 (>=3.2.3,<4.0.0); python_version >= \"2.7\" and python_version < \"2.8\"",
-      "typing (>=3.6,<4.0); python_version >= \"2.7\" and python_version < \"2.8\" or python_version >= \"3.4\" and python_version < \"3.5\""
-    ],
-    "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-    "summary": "Style preserving TOML library",
-    "version": "0.5.2",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "poetry",
-    "roles": [
-      {
-        "role": "Maintainer",
-        "user": "neersighted"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "4045c5f6848fbc93c38df2296a441f07",
-        "sha256": "dea8ff39e9e2170f1b2f465520482eec71e7909cfff53dcb076b585d50f8ccc8"
-      },
-      "downloads": -1,
-      "filename": "tomlkit-0.5.2-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "4045c5f6848fbc93c38df2296a441f07",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 116499,
-      "upload_time": "2018-11-09T17:09:28",
-      "upload_time_iso_8601": "2018-11-09T17:09:28.212157Z",
-      "url": "https://files.pythonhosted.org/packages/9b/ca/8b60a94c01ee655ffb81d11c11396cb6fff89459317aa1fe3e98ee80f055/tomlkit-0.5.2-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "7c31987ef6fba2cd64715cae27fade64",
-        "sha256": "4a226ccf11ee5a2e76bfc185747b54ee7718706aeb3aabb981327249dbe2b1d4"
-      },
-      "downloads": -1,
-      "filename": "tomlkit-0.5.2.tar.gz",
-      "has_sig": false,
-      "md5_digest": "7c31987ef6fba2cd64715cae27fade64",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 29813,
-      "upload_time": "2018-11-09T17:09:29",
-      "upload_time_iso_8601": "2018-11-09T17:09:29.709061Z",
-      "url": "https://files.pythonhosted.org/packages/f6/8c/c27d292cf7c0f04f0e1b5c75ab95dc328542ccbe9a809a1eada66c897bd2/tomlkit-0.5.2.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
