--- tests/repositories/fixtures/pypi.org/json/tomlkit/0.5.3.json
@@ -1,107 +1,0 @@
-{
-  "info": {
-    "author": "Sébastien Eustace",
-    "author_email": "sebastien@eustace.io",
-    "bugtrack_url": null,
-    "classifiers": [
-      "License :: OSI Approved :: MIT License",
-      "Programming Language :: Python :: 2",
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: 3.7"
-    ],
-    "description": "",
-    "description_content_type": "text/markdown",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/sdispater/tomlkit",
-    "keywords": "",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "Sébastien Eustace",
-    "maintainer_email": "sebastien@eustace.io",
-    "name": "tomlkit",
-    "package_url": "https://pypi.org/project/tomlkit/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/tomlkit/",
-    "project_urls": {
-      "Homepage": "https://github.com/sdispater/tomlkit",
-      "Repository": "https://github.com/sdispater/tomlkit"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/tomlkit/0.5.3/",
-    "requires_dist": [
-      "enum34 (>=1.1,<2.0); python_version >= \"2.7\" and python_version < \"2.8\"",
-      "functools32 (>=3.2.3,<4.0.0); python_version >= \"2.7\" and python_version < \"2.8\"",
-      "typing (>=3.6,<4.0); python_version >= \"2.7\" and python_version < \"2.8\" or python_version >= \"3.4\" and python_version < \"3.5\""
-    ],
-    "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-    "summary": "Style preserving TOML library",
-    "version": "0.5.3",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "poetry",
-    "roles": [
-      {
-        "role": "Maintainer",
-        "user": "neersighted"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "3a90c70a5067d5727110838094ab8674",
-        "sha256": "35f06da5835e85f149a4701d43e730adcc09f1b362e5fc2300d77bdd26280908"
-      },
-      "downloads": -1,
-      "filename": "tomlkit-0.5.3-py2.py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "3a90c70a5067d5727110838094ab8674",
-      "packagetype": "bdist_wheel",
-      "python_version": "py2.py3",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 116796,
-      "upload_time": "2018-11-19T20:05:37",
-      "upload_time_iso_8601": "2018-11-19T20:05:37.276181Z",
-      "url": "https://files.pythonhosted.org/packages/71/c6/06c014b92cc48270765d6a9418d82239b158d8a9b69e031b0e2c6598740b/tomlkit-0.5.3-py2.py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "cdbdc302a184d1f1e38d5e0810e3b212",
-        "sha256": "e2f785651609492c771d9887ccb2369d891d16595d2d97972e2cbe5e8fb3439f"
-      },
-      "downloads": -1,
-      "filename": "tomlkit-0.5.3.tar.gz",
-      "has_sig": false,
-      "md5_digest": "cdbdc302a184d1f1e38d5e0810e3b212",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
-      "size": 29864,
-      "upload_time": "2018-11-19T20:05:39",
-      "upload_time_iso_8601": "2018-11-19T20:05:39.200001Z",
-      "url": "https://files.pythonhosted.org/packages/f7/f7/bbd9213bfe76cb7821c897f9ed74877fd74993b4ca2fe9513eb5a31030f9/tomlkit-0.5.3.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
