--- tests/repositories/fixtures/pypi.org/json/twisted/18.9.0.json
@@ -1,75 +1,0 @@
-{
-  "info": {
-    "author": "Twisted Matrix Laboratories",
-    "author_email": "twisted-python@twistedmatrix.com",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Programming Language :: Python :: 2.7",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3.4",
-      "Programming Language :: Python :: 3.5",
-      "Programming Language :: Python :: 3.6",
-      "Programming Language :: Python :: 3.7"
-    ],
-    "description": "",
-    "description_content_type": "",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "http://twistedmatrix.com/",
-    "keywords": "",
-    "license": "MIT",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "Glyph Lefkowitz",
-    "maintainer_email": "glyph@twistedmatrix.com",
-    "name": "Twisted",
-    "package_url": "https://pypi.org/project/Twisted/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/Twisted/",
-    "project_urls": {
-      "Homepage": "http://twistedmatrix.com/"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/Twisted/18.9.0/",
-    "requires_dist": null,
-    "requires_python": "",
-    "summary": "An asynchronous networking framework written in Python",
-    "version": "18.9.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "twisted",
-    "roles": []
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "35ff4705ea90a76bf972ff3b229546ca",
-        "sha256": "4335327da58be11dd6e482ec6b85eb055bcc953a9570cd59e7840a2ce9419a8e"
-      },
-      "downloads": -1,
-      "filename": "Twisted-18.9.0.tar.bz2",
-      "has_sig": false,
-      "md5_digest": "35ff4705ea90a76bf972ff3b229546ca",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": null,
-      "size": 3088398,
-      "upload_time": "2018-10-15T09:11:22",
-      "upload_time_iso_8601": "2018-10-15T09:11:22.298247Z",
-      "url": "https://files.pythonhosted.org/packages/5d/0e/a72d85a55761c2c3ff1cb968143a2fd5f360220779ed90e0fadf4106d4f2/Twisted-18.9.0.tar.bz2",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
