--- tests/repositories/fixtures/pypi.org/json/wheel/0.40.0.json
@@ -1,117 +1,0 @@
-{
-  "info": {
-    "author": "",
-    "author_email": "Daniel Holth <dholth@fastmail.fm>",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Programming Language :: Python",
-      "Programming Language :: Python :: 3 :: Only",
-      "Programming Language :: Python :: 3.10",
-      "Programming Language :: Python :: 3.11",
-      "Programming Language :: Python :: 3.7",
-      "Programming Language :: Python :: 3.8",
-      "Programming Language :: Python :: 3.9",
-      "Topic :: System :: Archiving :: Packaging"
-    ],
-    "description": "",
-    "description_content_type": "text/x-rst",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "",
-    "keywords": "wheel,packaging",
-    "license": "",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "Alex Grönholm <alex.gronholm@nextday.fi>",
-    "name": "wheel",
-    "package_url": "https://pypi.org/project/wheel/",
-    "platform": null,
-    "project_url": "https://pypi.org/project/wheel/",
-    "project_urls": {
-      "Changelog": "https://wheel.readthedocs.io/en/stable/news.html",
-      "Documentation": "https://wheel.readthedocs.io/",
-      "Issue Tracker": "https://github.com/pypa/wheel/issues"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/wheel/0.40.0/",
-    "requires_dist": [
-      "pytest >= 6.0.0 ; extra == \"test\""
-    ],
-    "requires_python": ">=3.7",
-    "summary": "A built-package format for Python",
-    "version": "0.40.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": "pypa",
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "agronholm"
-      },
-      {
-        "role": "Owner",
-        "user": "joeforker"
-      },
-      {
-        "role": "Owner",
-        "user": "natefoo"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "517d39f133bd7b1ff17caf09784b7543",
-        "sha256": "d236b20e7cb522daf2390fa84c55eea81c5c30190f90f29ae2ca1ad8355bf247"
-      },
-      "downloads": -1,
-      "filename": "wheel-0.40.0-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "517d39f133bd7b1ff17caf09784b7543",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": ">=3.7",
-      "size": 64545,
-      "upload_time": "2023-03-14T15:10:00",
-      "upload_time_iso_8601": "2023-03-14T15:10:00.828550Z",
-      "url": "https://files.pythonhosted.org/packages/61/86/cc8d1ff2ca31a312a25a708c891cf9facbad4eae493b3872638db6785eb5/wheel-0.40.0-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "5f175a8d693f74878964d4fd29729ab7",
-        "sha256": "5cb7e75751aa82e1b7db3fd52f5a9d59e7b06905630bed135793295931528740"
-      },
-      "downloads": -1,
-      "filename": "wheel-0.40.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "5f175a8d693f74878964d4fd29729ab7",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=3.7",
-      "size": 96226,
-      "upload_time": "2023-03-14T15:10:02",
-      "upload_time_iso_8601": "2023-03-14T15:10:02.873691Z",
-      "url": "https://files.pythonhosted.org/packages/fc/ef/0335f7217dd1e8096a9e8383e1d472aa14717878ffe07c4772e68b6e8735/wheel-0.40.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
