--- tests/repositories/fixtures/pypi.org/json/zipp/3.5.0.json
@@ -1,112 +1,0 @@
-{
-  "info": {
-    "author": "Jason R. Coombs",
-    "author_email": "jaraco@jaraco.com",
-    "bugtrack_url": null,
-    "classifiers": [
-      "Development Status :: 5 - Production/Stable",
-      "Intended Audience :: Developers",
-      "License :: OSI Approved :: MIT License",
-      "Programming Language :: Python :: 3",
-      "Programming Language :: Python :: 3 :: Only"
-    ],
-    "description": "",
-    "description_content_type": "",
-    "docs_url": null,
-    "download_url": "",
-    "downloads": {
-      "last_day": -1,
-      "last_month": -1,
-      "last_week": -1
-    },
-    "dynamic": null,
-    "home_page": "https://github.com/jaraco/zipp",
-    "keywords": "",
-    "license": "",
-    "license_expression": null,
-    "license_files": null,
-    "maintainer": "",
-    "maintainer_email": "",
-    "name": "zipp",
-    "package_url": "https://pypi.org/project/zipp/",
-    "platform": "",
-    "project_url": "https://pypi.org/project/zipp/",
-    "project_urls": {
-      "Homepage": "https://github.com/jaraco/zipp"
-    },
-    "provides_extra": null,
-    "release_url": "https://pypi.org/project/zipp/3.5.0/",
-    "requires_dist": [
-      "sphinx ; extra == 'docs'",
-      "jaraco.packaging (>=8.2) ; extra == 'docs'",
-      "rst.linker (>=1.9) ; extra == 'docs'",
-      "pytest (>=4.6) ; extra == 'testing'",
-      "pytest-checkdocs (>=2.4) ; extra == 'testing'",
-      "pytest-flake8 ; extra == 'testing'",
-      "pytest-cov ; extra == 'testing'",
-      "pytest-enabler (>=1.0.1) ; extra == 'testing'",
-      "jaraco.itertools ; extra == 'testing'",
-      "func-timeout ; extra == 'testing'",
-      "pytest-black (>=0.3.7) ; (platform_python_implementation != \"PyPy\" and python_version < \"3.10\") and extra == 'testing'",
-      "pytest-mypy ; (platform_python_implementation != \"PyPy\" and python_version < \"3.10\") and extra == 'testing'"
-    ],
-    "requires_python": ">=3.6",
-    "summary": "Backport of pathlib-compatible object wrapper for zip files",
-    "version": "3.5.0",
-    "yanked": false,
-    "yanked_reason": null
-  },
-  "last_serial": 0,
-  "ownership": {
-    "organization": null,
-    "roles": [
-      {
-        "role": "Owner",
-        "user": "jaraco"
-      }
-    ]
-  },
-  "urls": [
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "da62cbd850ba32ba93817aab0f03a855",
-        "sha256": "ec508cd5a3ed3d126293cafb34611469f2aef7342f575c3b6e072b995dc9da1f"
-      },
-      "downloads": -1,
-      "filename": "zipp-3.5.0-py3-none-any.whl",
-      "has_sig": false,
-      "md5_digest": "da62cbd850ba32ba93817aab0f03a855",
-      "packagetype": "bdist_wheel",
-      "python_version": "py3",
-      "requires_python": ">=3.6",
-      "size": 5700,
-      "upload_time": "2021-07-02T23:51:45",
-      "upload_time_iso_8601": "2021-07-02T23:51:45.759726Z",
-      "url": "https://files.pythonhosted.org/packages/92/d9/89f433969fb8dc5b9cbdd4b4deb587720ec1aeb59a020cf15002b9593eef/zipp-3.5.0-py3-none-any.whl",
-      "yanked": false,
-      "yanked_reason": null
-    },
-    {
-      "comment_text": "",
-      "digests": {
-        "md5": "16bf2a24fae340052e8565c264d21092",
-        "sha256": "239d50954a15aa4b283023f18dc451ba811fb4d263f4dd6855642e4d1c80cc9f"
-      },
-      "downloads": -1,
-      "filename": "zipp-3.5.0.tar.gz",
-      "has_sig": false,
-      "md5_digest": "16bf2a24fae340052e8565c264d21092",
-      "packagetype": "sdist",
-      "python_version": "source",
-      "requires_python": ">=3.6",
-      "size": 13270,
-      "upload_time": "2021-07-02T23:51:47",
-      "upload_time_iso_8601": "2021-07-02T23:51:47.004396Z",
-      "url": "https://files.pythonhosted.org/packages/3a/9f/1d4b62cbe8d222539a84089eeab603d8e45ee1f897803a0ae0860400d6e7/zipp-3.5.0.tar.gz",
-      "yanked": false,
-      "yanked_reason": null
-    }
-  ],
-  "vulnerabilities": []
-}
